% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAlpha.R
\name{addAlpha}
\alias{addAlpha}
\title{Add an Alpha Channel to Colors}
\usage{
addAlpha(col, alpha = 0.5)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{alpha}{Alpha transparency values between 0 and 1.}
}
\value{
Character vector of hexadecimal colors with alpha channel.
}
\description{
Add transparency to colors.
}
\examples{

op <- par(no.readonly = TRUE)

addAlpha("yellow", 0.2)
addAlpha(2, 0.5)   # red

canvas(3)
polygon(circle(x=c(-1,0,1), y=c(1,-1,1), radius=2), 
        col=addAlpha(2:4, 0.4))

x <- rnorm(15000)
par(mfrow=c(1,2))
plot(x, type="p", col="blue" )
plot(x, type="p", col=addAlpha("blue", .2), 
     main="Better insight with alpha channel" )

par(op)

}
\seealso{
\code{\link{adjustcolor}}, \code{\link{colToOpaque}}

Other color.manipulation:
\code{\link[=darken]{darken()}},
\code{\link[=fade]{fade()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}}
}
\concept{color-manipulation}
\concept{color.manipulation}
