% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asHtml.R
\name{asHtml}
\alias{asHtml}
\title{Mark a character vector as HTML}
\usage{
asHtml(x)
}
\arguments{
\item{x}{a character vector, typically containing HTML markup}
}
\value{
\code{x}, with class \code{"html"} added
}
\description{
Tags a character vector with the S3 class \code{"html"} so that it
prints via \code{\link[=preview.html]{preview.html()}} as readable text instead of as
a raw character vector.
}
\examples{
asHtml("<b>bold</b>")

}
\seealso{
Other html:
\code{\link[=asFileLink]{asFileLink()}},
\code{\link[=asImg]{asImg()}},
\code{\link[=embedFile]{embedFile()}},
\code{\link[=escapeHtml]{escapeHtml()}},
\code{\link{htmlNotation}},
\code{\link{htmlSubscript}},
\code{\link[=toHtmlTable]{toHtmlTable()}}
}
\concept{formatting}
\concept{html}
