% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bezier.R
\name{bezier}
\alias{bezier}
\title{Bézier Geometry}
\usage{
bezier(x, y = NULL, nPoints = 100)
}
\arguments{
\item{x, y}{numeric vectors of control points.
Alternatively, \code{x} may be a list with components
\code{x} and \code{y}.}

\item{nPoints}{number of points used to approximate the curve.}
}
\value{
An object inheriting from class
\code{"bezierGeometry"}.
}
\description{
Create a Bézier curve from a set of control points.
}
\examples{
canvas(xlim = c(0, 1))

lines(
  bezier(
    x = c(0, 0.5, 1),
    y = c(0, 1, 0)
  ),
  col = "red",
  lwd = 2
)

}
\references{
Farin, G. (1993).
\emph{Curves and Surfaces for Computer Aided Geometric Design}.
Academic Press.
}
\seealso{
\link[graphics:lines]{graphics::lines}

Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=circle]{circle()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=polygonX]{polygonX()}},
\code{\link[=regPolygon]{regPolygon()}},
\code{\link[=ring]{ring()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
