% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{cmykToRgb}
\alias{cmykToRgb}
\title{Convert CMYK to RGB}
\usage{
cmykToRgb(col, maxColorValue = 1)
}
\arguments{
\item{col}{numeric CMYK matrix (columns C, M, Y, K).}

\item{maxColorValue}{maximum channel value.}
}
\value{
Numeric RGB matrix.
}
\description{
Convert CMYK colors to RGB.
}
\examples{
cmyk <- cmyToCmyk(rgbToCmy(c("#FF6347", "#4169E1"), maxColorValue = 255))
cmykToRgb(cmyk)

# back to hexadecimal colors
rgbToHex(t(cmykToRgb(cmyk) * 255))

}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
