% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{cmyToCmyk}
\alias{cmyToCmyk}
\title{Convert CMY to CMYK}
\usage{
cmyToCmyk(col)
}
\arguments{
\item{col}{numeric CMY matrix.}
}
\value{
Numeric CMYK matrix.
}
\description{
Convert CMY colors to CMYK.
}
\examples{
cmy <- rgbToCmy(c("#FF6347", "#4169E1"), maxColorValue = 255)
cmyToCmyk(cmy)

}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
