% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{colToHex}
\alias{colToHex}
\title{Convert R Colors to Hexadecimal Colors}
\usage{
colToHex(col, opacity = 1)
}
\arguments{
\item{col}{vector of valid R colors.}

\item{opacity}{opacity value between 0 and 1.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Convert any valid R color specification to hexadecimal colors.
}
\examples{
colToHex(c("tomato", "royalblue"))
colToHex("tomato", opacity = 0.5)

# color numbers refer to the current palette
colToHex(1:3)

}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
