% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_colors.R
\name{conv_cols}
\alias{conv_cols}
\alias{colToHex}
\alias{hexToRgb}
\alias{rgbToHex}
\alias{rgbToCmy}
\alias{cmykToRgb}
\alias{cmykToCmy}
\alias{cmyToCmyk}
\alias{rgbToCmyk}
\alias{colToHsv}
\title{Conversion Between RGB and CMYK}
\usage{
colToHex(col, alpha = 1)

hexToRgb(hex)

rgbToHex(col)

cmykToRgb(cyan, magenta, yellow, black, maxColorValue = 1)

rgbToCmy(col, maxColorValue = 1)

cmyToCmyk(col)

cmykToCmy(col)

colToHsv(col, alpha = FALSE)
}
\arguments{
\item{col}{the matrix of the color to be converted}

\item{alpha}{logical value indicating whether alpha channel (opacity) values
should be returned.}

\item{hex}{a color or a vector of colors specified as
hexadecimal string of the form "#RRGGBB" or "#RRGGBBAA"}

\item{cyan}{cyan values of the color(s) to be converted}

\item{magenta}{magenta values of the color(s) to be converted}

\item{yellow}{yellow values of the color(s) to be converted}

\item{black}{black values of the color(s) to be converted}

\item{maxColorValue}{the value for the color}
}
\value{
the converted value
}
\description{
These function convert colors between RGB and CMYK system.
}
\examples{

cmykToRgb(0.42, 45.23, 85.14, maxColorValue=100)

}
\seealso{
\code{\link{rgbToCol}}
}
