% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_coord.R
\name{conv_coord}
\alias{conv_coord}
\alias{polToCart}
\alias{cartToPol}
\alias{cartToSph}
\alias{sphToCart}
\title{Coordinate Transformations Cartesian/Polar/Spherical}
\usage{
polToCart(r, theta)

cartToPol(x, y)

cartToSph(x, y, z, up = TRUE)

sphToCart(r, theta, phi, up = TRUE)
}
\arguments{
\item{r}{a vector with the radius of the points.}

\item{theta}{a vector with the angle(s) of the points.}

\item{x, y, z}{vectors with the xy-coordianates to be transformed.}

\item{up}{logical. If set to \code{TRUE} (default) theta is measured from
x-y plane, else theta is measured from the z-axis.}

\item{phi}{a vector with the angle(s) of the points.}
}
\value{
\code{polToCart()} returns a list of x and y coordinates of the points.\cr
\code{cartToPol()} returns a list of r for the radius and theta for the angles of the
given points.
}
\description{
Transform cartesian into polar coordinates, resp. to spherical coordinates
and vice versa.
}
\details{
Angles are in radians, not degrees (i.e., a right angle is pi/2). Use
\code{\link{degToRad}} to convert, if you don't wanna do it by yourself.\cr
All parameters are recycled if necessary.
}
\note{
Based on code by Christian W. Hoffmann
}
\examples{

cartToPol(x=1, y=1)
cartToPol(x=c(1,2,3), y=c(1,1,1))
cartToPol(x=c(1,2,3), y=1)


polToCart(r=1, theta=pi/2)
polToCart(r=c(1,2,3), theta=pi/2)

polToCart(r=c(1,2,3), theta=pi/2)

cartToSph(x=1, y=2, z=3)   # r=3.741657, theta=0.930274, phi=1.107149

}
\keyword{arith}
\keyword{dplot}
