% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_rgb.R
\name{conv_rgb}
\alias{conv_rgb}
\alias{hexToCol}
\alias{rgbToCol}
\alias{colToRgb}
\alias{rgbToLong}
\alias{longToRgb}
\alias{colToGrey}
\alias{colToGray}
\title{Different Color Conversions}
\usage{
hexToCol(hex, method = "rgb", metric = "euclidean")

rgbToCol(col, method = "rgb", metric = "euclidean")

colToRgb(col, alpha = FALSE)

rgbToLong(col)

longToRgb(col)

colToGrey(col)

colToGray(col)
}
\arguments{
\item{hex}{a color or a vector of colors specified as
hexadecimal string of the form "#RRGGBB" or "#RRGGBBAA"}

\item{method}{character string specifying the color space to be used. Can be
\code{"rgb"} (default) or \code{"hsv"}.}

\item{metric}{character string specifying the metric to be used for
calculating distances between the colors.  Available options are
\code{"euclidean"} (default) and \code{"manhattan"}. Euclidean distances are
root sum-of-squares of differences, and manhattan distances are the sum of
absolute differences.}

\item{col}{vector of any of the three kind of R colors, i.e., either a color
name (an element of \code{\link{colors}}()), a hexadecimal string of the
form "#rrggbb" or "#rrggbbaa", or an integer i meaning palette()\verb{[i]}.
Non-string values are coerced to integer.}

\item{alpha}{logical value indicating whether alpha channel (opacity) values
should be returned.}
}
\value{
the name of the nearest found R color.
}
\description{
R color to RGB (red/green/blue) conversion is straight on. The specific
function is merely a wrapper to \code{\link{col2rgb}()}, defined in order to
follow this package's naming conventions.
}
\details{
Converting a RGB-color to a named R-Color means looking for a color in the
R-palette, which is nearest to the given RGB-color. This function uses the
minimum of squared distance (\code{"euclidean"}) or the minimum absolute
distance (\code{"manhattan"}) as proximity measure. \cr \code{rgbToLong()}
converts a RGB-color to a long integer in numeric format. \code{LongToRGB()}
does it the other way round.

It may not be clear from the start which method, rgb or hsv, yield the more
natural results. Trying and comparing is a recommended strategy. Moreover
the shortest numerical distance will not always be the best choice, when
comparing the colours visually.

\code{colToGrey()} converts colors to grey/grayscale
using the formula grey = 0.3\emph{red + 0.59}green + 0.11*blue.  This allows
you to see how your color plot will approximately
look when printed on a non-color printer or photocopied.
}
\note{
Converting to greyscales is based on code by Greg Snow.
}
\examples{

rgbToCol(matrix(c(162,42,42), nrow=3))

rgbToLong(matrix(c(162,42,42), nrow=3))

colToRgb("peachpuff")
colToRgb(c(blu = "royalblue", reddish = "tomato")) # names kept

colToRgb(1:8)

op <- par(no.readonly = TRUE)
par(mfcol=c(2,2))

tmp <- 1:3
names(tmp) <- c('red','green','blue')

barplot(tmp, col=c('red','green','blue'))
barplot(tmp, col=colToGrey(c('red','green','blue')))

barplot(tmp, col=c('red','#008100','#3636ff'))
barplot(tmp, col=colToGrey(c('red','#008100','#3636ff')))

par(op)

}
\seealso{
\code{\link{colToRgb}} and the other conversion functions

Other topic.colors: 
\code{\link{alpha}()},
\code{\link{colToOpaque}()},
\code{\link{contrastColor}()},
\code{\link{conv_cols}},
\code{\link{findColor}()}
}
\concept{color-conversion}
\concept{topic.colors}
