% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree-radians-conversion.R
\name{degree-radians-conversion}
\alias{degree-radians-conversion}
\alias{degToRad}
\alias{radToDeg}
\title{Convert Degrees to Radians and Vice Versa}
\usage{
degToRad(deg)

radToDeg(rad)
}
\arguments{
\item{deg}{a vector of angles in degrees.}

\item{rad}{a vector of angles in radians.}
}
\value{
degToRad returns a vector of the same length as \code{deg} with the
angles in radians.\cr radToDeg returns a vector of the same length as
\code{rad} with the angles in degrees.
}
\description{
Convert degrees to radians (and back again).
}
\examples{

degToRad(c(90,180,270))
radToDeg( c(0.5,1,2) * pi)

}
\seealso{
Other geometry.conversion:
\code{\link{coordinate-conversions}}
}
\concept{geometry}
\concept{geometry.conversion}
\concept{numerical-methods}
