% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawArc.R
\name{drawArc}
\alias{drawArc}
\title{Draw Elliptic Arc(s)}
\usage{
drawArc(
  x = 0,
  y = x,
  rx = 1,
  ry = rx,
  theta.1 = 0,
  theta.2 = 2 * pi,
  nv = 100,
  col = par("col"),
  lty = par("lty"),
  lwd = par("lwd"),
  plot = TRUE
)
}
\arguments{
\item{x, y}{a vector (or scalar) of xy-coordinates of the center(s) of the
arc(s).}

\item{rx}{a scalar or a vector giving the semi-major axis of the ellipse for
the arc(s)}

\item{ry}{a scalar or a vector giving the semi-minor axis of the ellipse for
the arc(s).  Default is radius.x which will result in a circle arc with
radius.x.}

\item{theta.1}{a scalar or a vector of starting angles in radians.}

\item{theta.2}{a scalar or a vector of ending angles in radians.}

\item{nv}{number of vertices used to plot the arc. Scalar or vector.}

\item{col}{color for the arc(s). Scalar or vector.}

\item{lty}{line type used for drawing.}

\item{lwd}{line width used for drawing.}

\item{plot}{logical. If \code{TRUE} the structure will be plotted. If
\code{FALSE} only the xy-points are calculated and returned.  Use this if
you want to combine several geometric structures to a single polygon.}
}
\value{
\code{drawArc} invisibly returns a list of the calculated
coordinates for all shapes.
}
\description{
Draw one or more elliptic (or circular) arcs from \code{theta.1} to
\code{theta.2} on an existing plot using classic graphics.
}
\details{
All parameters are recycled if necessary. \cr Be sure to use an aspect ratio
of 1 as shown in the example to avoid distortion.
}
\examples{

curve(sin(x), 0, pi, col="blue", asp=1)
drawArc(x = pi/2, y = 0, rx = 1, theta.1 = pi/4, theta.2 = 3*pi/4, col="red")

}
\seealso{
\code{\link{polygon}}

Other plot.geometry:
\code{\link[=abcCoords]{abcCoords()}},
\code{\link[=canvas]{canvas()}},
\code{\link[=drawBezier]{drawBezier()}},
\code{\link[=drawCircle]{drawCircle()}},
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=drawRegPolygon]{drawRegPolygon()}},
\code{\link[=polarGrid]{polarGrid()}}
}
\concept{geometry}
\concept{graphics}
\concept{plot.geometry}
