% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEllipse.R
\name{drawEllipse}
\alias{drawEllipse}
\title{Draw an Ellipse}
\usage{
drawEllipse(
  x = 0,
  y = x,
  radius.x = 1,
  radius.y = 0.5,
  rot = 0,
  nv = 100,
  border = par("fg"),
  col = par("bg"),
  lty = par("lty"),
  lwd = par("lwd"),
  plot = TRUE
)
}
\arguments{
\item{x, y}{the x and y co-ordinates for the centre(s) of the ellipse(s).}

\item{radius.x}{a scalar or a vector giving the semi-major axis of the
ellipse.}

\item{radius.y}{a scalar or a vector giving the semi-minor axis of the
ellipse.}

\item{rot}{angle of rotation in radians.}

\item{nv}{number of vertices to draw the ellipses.}

\item{border}{color for borders. The default is \code{par("fg")}. Use
\code{border = NA} to omit borders.}

\item{col}{color(s) to fill or shade the annulus sector with. The default
\code{NA} (or also \code{NULL}) means do not fill (say draw transparent).}

\item{lty}{line type for borders and shading; defaults to \code{"solid"}.}

\item{lwd}{line width for borders and shading.}

\item{plot}{logical. If \code{TRUE} the structure will be plotted. If
\code{FALSE} only the points are calculated and returned. Use this if you
want to combine several geometric structures to a single polygon.}
}
\value{
The function invisibly returns a list of the calculated coordinates
for all shapes.
}
\description{
Draw one or several ellipses on an existing plot.
}
\details{
Use \code{\link{degToRad}} if you want to define rotation angle in degrees.
}
\examples{

op <- par(no.readonly = TRUE)
par(mfrow=c(1,2))

canvas()
drawEllipse(rot = c(1:3) * pi/3, col=alpha(c("blue","red","green"), 0.5) )


plot(cars)
m <- var(cars)
eig <- eigen(m)
eig.val <- sqrt(eig$values)
eig.vec <- eig$vectors

drawEllipse(x=mean(cars$speed), y=mean(cars$dist), 
            radius.x=eig.val[1] , radius.y=eig.val[2], 
            rot=acos(eig.vec[1,1]), border="blue", lwd=3)
            
par(op)

}
\seealso{
\code{\link{polygon}}, \code{\link{drawRegPolygon}},
\code{\link{drawCircle}}, \code{\link{drawArc}}

Other plot.geometry: 
\code{\link{canvas}()},
\code{\link{drawArc}()},
\code{\link{drawBezier}()},
\code{\link{drawCircle}()},
\code{\link{drawRegPolygon}()},
\code{\link{polarGrid}()}
}
\concept{geometry}
\concept{graphics}
\concept{plot.geometry}
