% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmCI.R
\name{fmCI}
\alias{fmCI}
\title{Format Confidence Intervals}
\usage{
fmCI(x, template = NULL, ...)
}
\arguments{
\item{x}{the numerical values, given in the order in which they are used in
the template.}

\item{template}{character string as template for the desired format.
\%s are the placeholders for the numerical values. Default
is \verb{`"\%s [\%s, \%s]"`} for \verb{<est> [<lci>, <uci>]}.}

\item{\dots}{the dots are passed on to the \code{\link[=fm]{fm()}} function.}
}
\value{
a formatted string
}
\description{
Format confidence intervals using a flexible template.
}
\examples{

x <- c(est=2.1, lci=1.5, uci=3.8)

# default template
fmCI(x)
# user defined template (note the double percent code)
fmCI(x, template="\%s (95\%\%-CI \%s-\%s)", digits=1)

# in higher datastructures
tapply(warpbreaks$breaks, warpbreaks$wool, 
       function(x) fmCI(c(mean(x), t.test(x)$conf.int),
                        digits=1))

}
\seealso{
\code{\link[=fm]{fm()}}

Other format:
\code{\link[=convUnit]{convUnit()}},
\code{\link[=fm]{fm()}},
\code{\link[=print.Unit]{print.Unit()}},
\code{\link[=style]{style()}},
\code{\link[=unit]{unit()}}
}
\concept{confidence-interval}
\concept{format}
\concept{formatting}
