% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-notation.R
\name{\%_\%}
\alias{\%_\%}
\title{Subscript notation}
\usage{
x \%_\% i
}
\arguments{
\item{x}{a character vector (the base symbol)}

\item{i}{a character vector (the subscript), recycled against \code{x}}
}
\value{
a character vector: \code{x} followed by \verb{<sub>i</sub>}
}
\description{
Infix operator producing an HTML subscript, e.g. for indexed symbols
such as \eqn{x_i}.
}
\examples{
"x" \%_\% "i"  # -> "x<sub>i</sub>"

}
