% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grayscale.R
\name{grayScale}
\alias{grayScale}
\title{Convert Colors to grayScale}
\usage{
grayScale(col)
}
\arguments{
\item{col}{vector of valid R colors.}
}
\value{
Character vector of grayScale colors.
}
\description{
Convert colors to grayScale using luminance weighting.
}
\details{
grayScale conversion uses the standard luminance approximation:
\deqn{0.3 R + 0.59 G + 0.11 B}
}
\examples{
op <- par(no.readonly = TRUE)
par(mfcol=c(2,2))

tmp <- 1:3
names(tmp) <- c('red','green','blue')

barplot(tmp, col=c('red','green','blue'))
barplot(tmp, col=grayScale(c('red','green','blue')))

barplot(tmp, col=c('red','#008100','#3636ff'))
barplot(tmp, col=grayScale(c('red','#008100','#3636ff')))

par(op)

}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
