% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{hexToCol}
\alias{hexToCol}
\title{Convert Hex Colors to Named R Colors}
\usage{
hexToCol(col, method = c("rgb", "hsv"), metric = c("euclidean", "manhattan"))
}
\arguments{
\item{col}{character vector of hexadecimal colors.}

\item{method}{color space used for matching.}

\item{metric}{distance metric.}
}
\value{
Character vector of named R colors.
}
\description{
Convert hexadecimal colors to the nearest named R colors.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
