% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{hexToRgb}
\alias{hexToRgb}
\title{Convert Hex Colors to RGB}
\usage{
hexToRgb(col)
}
\arguments{
\item{col}{character vector of hexadecimal colors.}
}
\value{
Integer matrix with RGB rows.
}
\description{
Convert hexadecimal color strings to an RGB matrix.
}
\examples{
hexToRgb(c("#A52A2A", "#FFDAB9"))

# an alpha channel is kept
hexToRgb("#A52A2A80")

}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
