% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{longToRgb}
\alias{longToRgb}
\title{Convert Long Integers to RGB}
\usage{
longToRgb(col)
}
\arguments{
\item{col}{integer vector.}
}
\value{
RGB matrix.
}
\description{
Decode long integers into RGB values.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
