% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixColors.R
\name{mixColors}
\alias{mixColors}
\title{Mix Colors}
\usage{
mixColors(col1, col2, weight = 0.5)
}
\arguments{
\item{col1}{First vector of colors.}

\item{col2}{Second vector of colors.}

\item{weight}{Numeric value between 0 and 1 specifying the
contribution of \code{col2}. A value of 0 returns \code{col1},
while 1 returns \code{col2}.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Mix two sets of colors in RGB space.
}
\details{
Colors are mixed linearly in RGB space:
\deqn{
  x_{mix} = (1 - weight) x_1 + weight x_2
}

All arguments are recycled as necessary.
}
\seealso{
\code{\link{lighten}},
\code{\link{darken}}

Other color.manipulation:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=darken]{darken()}},
\code{\link[=fade]{fade()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=lighten]{lighten()}}
}
\concept{color-manipulation}
\concept{color.manipulation}
