% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html-notation.R
\name{notation}
\alias{notation}
\alias{htmlHat}
\alias{htmlBar}
\title{HTML notation for hat and bar diacritics}
\usage{
htmlHat(x)

htmlBar(x)
}
\arguments{
\item{x}{a character vector, typically a variable name or symbol}
}
\value{
a character vector with the diacritic's HTML entity appended
}
\description{
Append the HTML entity for a circumflex (\dQuote{hat}, as in an
estimator \eqn{\hat{x}}) or macron (\dQuote{bar}, as in a sample mean
\eqn{\bar{x}}) to a label.
}
\examples{
htmlHat("p")  # -> "p&#770;"   (renders as p with a circumflex)
htmlBar("x")  # -> "x&#772;"   (renders as x with a macron)

}
