% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Desc.ts.R
\name{plot.Desc.ts}
\alias{plot.Desc.ts}
\title{Plot frequency tables with bar or dot representation}
\usage{
\method{plot}{Desc.ts}(x, ...)
}
\arguments{
\item{x}{An object of class \code{Desc.table}.}

\item{...}{Graphical parameters passed to \code{par()}, such as
\code{cex}, \code{las}, \code{mar}, or \code{oma}.}
}
\description{
Creates a ACF plot for objects produced by \code{Desc.ts()}.
Two mosaicplots are placed side by side
}
\details{
Graphical parameters supplied via \code{...} are applied globally using
\code{par()} and affect all subplots consistently.
}
\seealso{
Other topic.graphics: 
\code{\link{axisBreak}()},
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{canvas}()},
\code{\link{colorLegend}()},
\code{\link{errBars}()},
\code{\link{plot.Desc.factor}()},
\code{\link{plot.Desc.logical}()},
\code{\link{plot.Desc.table}()},
\code{\link{plot.Lc}()},
\code{\link{plot.PostHocTest}()},
\code{\link{plotArea}()},
\code{\link{plotBag}()},
\code{\link{plotBar}()},
\code{\link{plotBubble}()},
\code{\link{plotCirc}()},
\code{\link{plotCor}()},
\code{\link{plotDens}()},
\code{\link{plotDens2D}()},
\code{\link{plotDot}()},
\code{\link{plotECDF}()},
\code{\link{plotFdist}()},
\code{\link{plotFun}()},
\code{\link{plotLines}()},
\code{\link{plotMiss}()},
\code{\link{plotMosaic}()},
\code{\link{plotProbDist}()},
\code{\link{plotQQ}()},
\code{\link{plotRidge}()},
\code{\link{plotTernary}()},
\code{\link{plotTimeSeries}()},
\code{\link{plotTreemap}()},
\code{\link{plotViolin}()},
\code{\link{plotWeb}()}
}
\concept{base-graphics}
\concept{plotting}
\concept{topic.graphics}
