% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMiss.R
\name{plotMiss}
\alias{plotMiss}
\title{Plot Missing Data}
\usage{
plotMiss(
  x,
  col = "deeppink4",
  bg = fade("navajowhite3", 0.3),
  clust = FALSE,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{a data.frame to be analysed.}

\item{col}{the colour of the missings.}

\item{bg}{the background colour of the plot.}

\item{clust}{logical, defining if the missings should be clustered. Default
is \code{FALSE}.}

\item{main}{the main title.}

\item{\dots}{the dots are passed to \code{\link[=plot]{plot()}}.}
}
\value{
if clust is set to TRUE, the new order will be returned invisibly.
}
\description{
Takes a data frame and displays the location of missing data. The missings
can be clustered and be displayed together.
}
\details{
A graphical display of the position of the missings can be help to detect
dependencies or patterns within the missings.
}
\note{
Following an idea of Henk Harmsen \href{mailto:henk@carbonmetrics.com}{henk@carbonmetrics.com}
}
\examples{

plotMiss(airquality, main="Missing data (in orignal order)")
plotMiss(airquality, main="Missing data (clustered)", clust=TRUE)

}
\seealso{
\code{\link[=hclust]{hclust()}}, \code{\link[bedrock:countCompCases]{bedrock::countCompCases()}}

Other plot.special:
\code{\link[=plotBinaryTree]{plotBinaryTree()}},
\code{\link[=plotCirc]{plotCirc()}},
\code{\link[=plotLift]{plotLift()}},
\code{\link[=plotPolar]{plotPolar()}},
\code{\link[=plotPropCI]{plotPropCI()}},
\code{\link[=plotTernary]{plotTernary()}},
\code{\link[=plotTimeSeries]{plotTimeSeries()}},
\code{\link[=plotTreemap]{plotTreemap()}},
\code{\link[=plotWeb]{plotWeb()}}
}
\concept{missing-value}
\concept{plot.special}
