% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQQ.R
\name{plotQQ}
\alias{plotQQ}
\title{QQ-Plot for Any Distribution}
\usage{
plotQQ(
  x,
  qFun = qnorm,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  datax = FALSE,
  add = FALSE,
  cex = .useTheme,
  grid = .useTheme,
  box = .useTheme,
  cband = list(conf.level = 0.95),
  qqline = TRUE,
  stamp = .useTheme,
  ...
)
}
\arguments{
\item{x}{the data sample}

\item{qFun}{the quantile function of the assumed distribution. Can either
be given as simple function name or defined as own function using the
required arguments. Default is \code{qnorm()}. See examples.}

\item{main}{the main title for the plot. This will be "Q-Q-Plot" by default}

\item{xlab}{the xlab for the plot}

\item{ylab}{the ylab for the plot}

\item{datax}{logical. Should data values be on the x-axis? Default is
\code{FALSE}.}

\item{add}{logical specifying if the points should be added to an already
existing plot; defaults to \code{FALSE}.}

\item{cex}{character expansion factor for the points. \code{.useTheme}
(default) resolves to \code{getTheme()$points$cex}.}

\item{grid}{controls drawing of the background grid. \code{.useTheme}
(default) follows the active theme (\code{getTheme()$grid}).
\code{TRUE}/\code{FALSE}/\code{NA}, or a named list, as for
\code{\link[graphics:grid]{graphics::grid()}}.}

\item{box}{controls drawing of the plot box. \code{.useTheme} (default)
resolves to \code{getTheme()$box}. \code{TRUE}/\code{FALSE}/\code{NA},
or a named list, as for \code{\link[graphics:box]{graphics::box()}}.}

\item{cband}{controls the confidence band. \code{FALSE}, \code{NULL}, or
\code{NA} suppress it. A named list configures it; its
\code{conf.level} element (default \code{0.95}) controls the
confidence level of the pointwise Kolmogorov-Smirnov-based band
(\code{conf.level} is consumed before the band is drawn and never
forwarded as a graphical parameter). Other list elements (e.g.
\code{col}, \code{border}) configure the band's appearance.}

\item{qqline}{arguments for the qqline. This will be estimated as a
line through the 25\\% and 75\\% quantiles by default, which is the same
procedure as \code{\link[=qqline]{qqline()}} does for normal distribution (instead of
set it to \verb{abline(a = 0, b = 1))}. The quantiles can however be
overwritten by setting the argument \code{probs} to some user defined
values. Also the method for calculating the quantiles can be defined
(default is 7, see \code{\link[=quantile]{quantile()}}). The line defaults are set to
\code{col = par("fg")}, \code{lwd = par("lwd")} and \code{lty = par("lty")}.
No line will be plotted if \code{args.qqline} is set to \code{NA}.}

\item{stamp}{controls the corner stamp. \code{.useTheme} (default)
resolves to \code{getTheme()$stamp}. \code{TRUE}/\code{FALSE}/\code{NULL},
a string, or a named list of arguments for \code{stamp()} (e.g.
\code{list(text = "...", las = 2)}).}

\item{\dots}{the dots are passed to the plot function.}
}
\value{
No return value (invisible \code{NULL}), called for its side effect of
drawing the plot.
}
\description{
Create a QQ-plot for a variable of any distribution. The assumed underlying
distribution can be defined as a function of f(p), including all required
parameters. Confidence bands are provided by default.
}
\details{
The function generates a sequence of points between 0 and 1 and transforms
those into quantiles by means of the defined assumed distribution.
}
\note{
The code is inspired by the tip 10.22 "Creating other
Quantile-Quantile plots" from R Cookbook and based on R-Core code from the
function \code{qqline}. The calculation of confidence bands are rewritten
based on an algorithm published in the package
\code{BoutrosLab.plotting.general}.

Based on code by Ying Wu
}
\examples{

y <- rexp(100, 1/10)
plotQQ(y, function(p) qexp(p, rate=1/10))

w <- rweibull(100, shape=2)
plotQQ(w, qFun = function(p) qweibull(p, shape=4))

z <- rchisq(100, df=5)
plotQQ(z, function(p) qchisq(p, df=5),
       qqline=list(col=2, probs=c(0.1, 0.6)),
       main=expression("Q-Q plot for" ~~ {chi^2}[nu == 3]))
abline(0,1)

# 90\% confidence band instead of the 95\% default
plotQQ(y, function(p) qexp(p, rate=1/10), cband = list(conf.level = 0.90))

# add 5 random sets
for(i in 1:5){
  z <- rchisq(100, df=5)
  plotQQ(z, function(p) qchisq(p, df=5), add=TRUE, qqline = FALSE,
         col="grey", lty="dotted")
}

}
\references{
Teetor, P. (2011) \emph{R Cookbook}. O'Reilly, pp. 254-255.
}
\seealso{
\link[stats:qqnorm]{stats::qqnorm}, \link[stats:qqline]{stats::qqline}, \link[stats:qqplot]{stats::qqplot},
\link{theme}, \link{lines.loess}

Other plot.univariate:
\code{\link[=plotArea]{plotArea()}},
\code{\link[=plotBar]{plotBar()}},
\code{\link[=plotBox]{plotBox()}},
\code{\link[=plotCatDist]{plotCatDist()}},
\code{\link[=plotDens]{plotDens()}},
\code{\link[=plotDensBox]{plotDensBox()}},
\code{\link[=plotDot]{plotDot()}},
\code{\link[=plotECDF]{plotECDF()}},
\code{\link[=plotFreqDist]{plotFreqDist()}},
\code{\link[=plotLines]{plotLines()}},
\code{\link[=plotViolin]{plotViolin()}}
}
\concept{distribution-summary}
\concept{normality-test}
\concept{plot.univariate}
