% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Combined Plot of a Time Series and Its ACF and PACF}
\usage{
plotTimeSeries(
  x,
  main = NULL,
  ylab = NULL,
  maxLag = 10 * log10(length(x)),
  stamp = .useTheme,
  ...
)
}
\arguments{
\item{x}{univariate time series.}

\item{main}{an overall title for the plot}

\item{ylab}{a title for the y axis: see \code{\link[=title]{title()}}.}

\item{maxLag}{integer. Defines the number of lags to be displayed. The
default is 10 * log10(length(series)).}

\item{stamp}{controls the corner stamp. \code{.useTheme} (default)
resolves to \code{getTheme()$stamp}. \code{TRUE}/\code{FALSE}/
\code{NULL}, a string, or a named list for \code{\link[=stamp]{stamp()}}.}

\item{\dots}{the dots are passed to the plot command.}
}
\description{
Combined plot of a time Series and its autocorrelation and partial
autocorrelation
}
\details{
plotTimeSeries plots a combination of the time series and its autocorrelation and
partial autocorrelation.
}
\note{
Rewritten based on ideas of M.Huerzeler
}
\examples{

plotTimeSeries(AirPassengers)
}
\seealso{
\code{\link[=ts]{ts()}}

Other plot.special:
\code{\link[=plotBinaryTree]{plotBinaryTree()}},
\code{\link[=plotCirc]{plotCirc()}},
\code{\link[=plotLift]{plotLift()}},
\code{\link[=plotMiss]{plotMiss()}},
\code{\link[=plotPolar]{plotPolar()}},
\code{\link[=plotPropCI]{plotPropCI()}},
\code{\link[=plotTernary]{plotTernary()}},
\code{\link[=plotTreemap]{plotTreemap()}},
\code{\link[=plotWeb]{plotWeb()}}
}
\concept{plot.special}
\concept{time-series}
