% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-methods.R
\name{polygonX}
\alias{polygonX}
\alias{polygonX.ringGeometry}
\alias{polygonX.polygonGeometry}
\alias{polygonX.geometryCollection}
\alias{polygonX.default}
\title{Draw Polygonal Geometries}
\usage{
polygonX(x, ...)

\method{polygonX}{default}(
  x,
  y = NULL,
  density = NULL,
  angle = 45,
  border = NULL,
  col = NA,
  lty = par("lty"),
  ...,
  fillOddEven = FALSE
)

\method{polygonX}{ringGeometry}(x, rule = "evenodd", ...)

\method{polygonX}{polygonGeometry}(x, ...)
}
\arguments{
\item{x}{an object to be drawn.}

\item{...}{further arguments passed to the corresponding method.}

\item{y}{numeric vector of y-coordinates.}

\item{density}{density of shading lines.}

\item{angle}{angle of shading lines in degrees.}

\item{border}{border colour.}

\item{col}{fill colour.}

\item{lty}{line type.}

\item{fillOddEven}{logical; should the odd-even rule be used for filling?}

\item{rule}{character string specifying the filling rule passed to
\code{\link[graphics:polypath]{graphics::polypath()}}. One of \code{"evenodd"} or
\code{"winding"}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Generic function for drawing polygon-based geometry objects.
This function extends graphics::polygon() with support for
geometry objects such as circle(), ellipse(), regPolygon() and ring()
and further remains fully compatible with its original interface.
#'
For ordinary coordinate vectors the call is forwarded to
\code{\link[graphics:polygon]{graphics::polygon()}}. Geometry objects such as
\code{\link[=circle]{circle()}}, \code{\link[=ellipse]{ellipse()}}, \code{\link[=regPolygon]{regPolygon()}}
and \code{\link[=ring]{ring()}} are dispatched to specialised methods.
}
\examples{
canvas()

polygonX(
  circle(radius = 1),
  col = "lightblue"
)

polygonX(
  regPolygon(
    radius = 0.7,
    nVertices = 5
  ),
  border = "red"
)

}
\seealso{
Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=circle]{circle()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=regPolygon]{regPolygon()}},
\code{\link[=ring]{ring()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
