% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.R
\name{print.Unit}
\alias{print.Unit}
\title{Print Object with Unit}
\usage{
\method{print}{Unit}(x, ...)
}
\arguments{
\item{x}{An object with class \code{"Unit"}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
S3 method for printing objects with a \code{"Unit"} class.
Displays the value along with its associated unit.
}
\examples{
x <- 10
unit(x) <- "m"
class(x) <- "Unit"
print(x)

}
\seealso{
Other format:
\code{\link[=convUnit]{convUnit()}},
\code{\link[=fm]{fm()}},
\code{\link[=fmCI]{fmCI()}},
\code{\link[=style]{style()}},
\code{\link[=unit]{unit()}}
}
\concept{annotation}
\concept{format}
