% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{regPolygon}
\alias{regPolygon}
\title{Regular Polygon Geometry}
\usage{
regPolygon(x = 0, y = 0, radius = 1, nVertices = 6, startAngle = 0)
}
\arguments{
\item{x, y}{centre coordinates.}

\item{radius}{circumradius.}

\item{nVertices}{number of vertices.}

\item{startAngle}{staring angle in radians.}
}
\value{
An object inheriting from class \code{"regPolygonGeometry"}.
}
\description{
Create a regular polygon.
}
\examples{
canvas(xlim = c(-2, 2))

# a hexagon is the default
polygonX(regPolygon(radius = 1.8), col = "lightblue")

# a square, once standing on a vertex and once on an edge
polygonX(regPolygon(radius = 1.2, nVertices = 4), border = "red")
polygonX(regPolygon(radius = 1.2, nVertices = 4, startAngle = pi/4))

# a triangle pointing upwards
polygonX(regPolygon(radius = 0.5, nVertices = 3, startAngle = pi/2),
         col = "royalblue")

}
\seealso{
Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=circle]{circle()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=polygonX]{polygonX()}},
\code{\link[=ring]{ring()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
