% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{rgbToCol}
\alias{rgbToCol}
\title{Convert RGB Colors to the Nearest Named R Color}
\usage{
rgbToCol(col, method = c("rgb", "hsv"), metric = c("euclidean", "manhattan"))
}
\arguments{
\item{col}{RGB matrix or hexadecimal colors.}

\item{method}{color space used for matching.}

\item{metric}{distance metric.}
}
\value{
Character vector of named R colors.
}
\description{
Match RGB colors to the nearest named R color.
}
\examples{
rgbToCol(matrix(c(162, 42, 42), nrow = 3))
rgbToCol(c("#A22A2A", "#4169E0"))

# nearest color in HSV space
rgbToCol(c("#A22A2A", "#4169E0"), method = "hsv", metric = "manhattan")

}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
