% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{rgbToLong}
\alias{rgbToLong}
\title{Convert RGB to Long Integers}
\usage{
rgbToLong(col)
}
\arguments{
\item{col}{RGB matrix.}
}
\value{
Integer vector.
}
\description{
Encode RGB colors as long integers.
}
\examples{
rgbToLong(colToRgb(c("red", "green", "blue")))

}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
