% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{ring}
\alias{ring}
\title{Ring Geometry}
\usage{
ring(
  x = 0,
  y = 0,
  innerRadius = 0.5,
  outerRadius = 1,
  startAngle = 0,
  endAngle = 2 * pi,
  nPoints = 100
)
}
\arguments{
\item{x, y}{centre coordinates.}

\item{innerRadius}{radius of the inner boundary.}

\item{outerRadius}{radius of the outer boundary.}

\item{startAngle, endAngle}{start and end angle in radians.}

\item{nPoints}{number of points used for each boundary.}
}
\value{
An object inheriting from class \code{"ringGeometry"} or a
\code{"geometryCollection"}.
}
\description{
Create one or more rings or ring segments.
}
\examples{
canvas(xlim = c(-2, 2))

# a full ring
polygonX(ring(innerRadius = 1.4, outerRadius = 1.8), col = "lightblue")

# two ring segments at once
r <- ring(innerRadius = c(0.3, 0.8), outerRadius = c(0.7, 1.2),
          startAngle = c(0, pi), endAngle = c(pi/2, 3*pi/2))
polygonX(r, col = c("tomato", "royalblue"))

}
\seealso{
Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=circle]{circle()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=polygonX]{polygonX()}},
\code{\link[=regPolygon]{regPolygon()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
