% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadOut.R
\name{spreadOut}
\alias{spreadOut}
\title{Spread Out a Vector of Numbers To a Minimum Interval}
\usage{
spreadOut(x, mindist = NULL, cex = 1)
}
\arguments{
\item{x}{a numeric vector which may contain \code{NA}s.}

\item{mindist}{the minimum interval between any two values. If this is left
to \code{NULL} (default) the function will check if a plot is open and then
use 90\%\% of \code{\link[=strheight]{strheight()}}.}

\item{cex}{numeric character expansion factor; multiplied by
\verb{[par]("cex")} yields the final character size; the default
\code{NULL} is equivalent to \code{1}.}
}
\value{
On success, the spread out values. If there are less than two valid
values, the original vector is returned.
}
\description{
Spread the numbers of a vector so that there is a minimum interval between
any two numbers (in ascending or descending order). This is helpful when we
want to place textboxes on a plot and ensure, that they do not mutually
overlap.
}
\details{
\code{spreadOut()} starts at or near the middle of the vector and increases
the intervals between the ordered values. \code{NA}s are preserved.
\code{spreadOut()} first tries to spread groups of values with intervals
less than \code{mindist} out neatly away from the mean of the group. If this
doesn't entirely succeed, a second pass that forces values away from the
middle is performed.

\code{spreadOut()} can also be used to avoid overplotting of axis tick
labels where they may be close together.
}
\note{
This function is based on \code{plotrix::spreadOut()} and has been
integrated here with some minor changes.

Based on code by Jim Lemon \href{mailto:jim@bitwrit.com.au}{jim@bitwrit.com.au}
}
\examples{

spreadOut(c(1, 3, 3, 3, 3, 5), 0.2)
spreadOut(c(1, 2.5, 2.5, 3.5, 3.5, 5), 0.2)
spreadOut(c(5, 2.5, 2.5, NA, 3.5, 1, 3.5, NA), 0.2)

# this will almost always invoke the brute force second pass
spreadOut(rnorm(10), 0.5)


}
\seealso{
\code{\link[=strheight]{strheight()}}

Other graphics.layout:
\code{\link[=abcCoords]{abcCoords()}},
\code{\link{axTicks}},
\code{\link[=axisBreak]{axisBreak()}},
\code{\link[=isValidPlotRegion]{isValidPlotRegion()}},
\code{\link[=lineToUser]{lineToUser()}},
\code{\link[=mar]{mar()}},
\code{\link[=plotFacet]{plotFacet()}}
}
\concept{graphics.layout}
\concept{label}
\concept{numerical-methods}
