% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strAlign.R
\name{strAlign}
\alias{strAlign}
\title{Align Strings}
\usage{
strAlign(x, sep = "\\\\r")
}
\arguments{
\item{x}{a character vector}

\item{sep}{a character specifying the alignment mode:
\itemize{
\item \code{"\\\\l"}: left alignment
\item \code{"\\\\r"}: right alignment (default)
\item \code{"\\\\c"}: centered alignment
\item any other character: align at the first occurrence of this separator
}}
}
\value{
A character vector of aligned strings.
}
\description{
Aligns a character vector to the left, right, center, or with respect
to the first occurrence of a specified separator.
}
\details{
Alignment is achieved by padding strings with spaces so that all elements
have equal width. This is mainly useful for monospaced output.

For left, right, and center alignment, strings are padded to the maximum
width using spaces.

If a separator is provided, strings are aligned such that the first
occurrence of the separator is vertically aligned. If a string does not
contain the separator, it is treated as if the separator occurred at the end.

This function uses Unicode-aware string handling via the \pkg{stringi}
package.
}
\examples{
x <- c("here", "there", "everywhere")

# right align (default)
strAlign(x)

# left align
strAlign(x, "\\\\l")

# center align
strAlign(x, "\\\\c")

# align at decimal separator
z <- c("6.0", "45.12", "784")
strAlign(z, ".")

}
\seealso{
\code{\link[stringi:stri_pad]{stringi::stri_pad()}},
\code{\link[stringi:stri_sub]{stringi::stri_sub()}},
\code{\link[stringi:stri_trim_right]{stringi::stri_trim_right()}}

\link{string-overview} for an overview of all string utilities in pharos.
}
\concept{formatting}
\concept{string-manipulation}
