% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strChop.R
\name{strChop}
\alias{strChop}
\title{Split a String into a Number of Sections of Defined Length}
\usage{
strChop(x, len = NULL, pos = NULL)
}
\arguments{
\item{x}{the string to be cut in pieces}

\item{len}{a vector with the lengths of the pieces}

\item{pos}{a vector of cutting positions. Will be ignored when \code{len}
has been defined.}
}
\value{
a vector with the parts of the string.
}
\description{
Splitting a string into a number of sections of defined length is needed,
when we want to split a table given as a number of lines without separator
into columns. The cutting points can either be defined by the lengths of the
sections or directly by position.
}
\details{
If length is going over the end of the string the last part will be
returned, so if the rest of the string is needed, it's possible to simply
enter a big number as last partlength.

\code{len} and \code{pos} can't be defined simultaneously, only
alternatively.

Typical usages are \preformatted{ strChop(x, len) strChop(x, pos) }
}
\examples{

x <- paste(letters, collapse="")
strChop(x=x, len = c(3,5,2))

# and with the rest integrated
strChop(x=x, len = c(3, 5, 2, nchar(x)))

# cutpoints at 5th and 10th position
strChop(x=x, pos=c(5, 10))

}
\seealso{
\code{\link[=strLeft]{strLeft()}},
\code{\link[=substr]{substr()}}

\link{string-overview} for an overview of all string utilities in pharos.
}
\concept{reshape}
\concept{string-manipulation}
