% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strCountWords.R
\name{strCountWords}
\alias{strCountWords}
\title{Count Words in Strings}
\usage{
strCountWords(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
An integer vector giving the number of words in each element of
\code{x}.
}
\description{
Counts the number of words in each element of a character vector.
}
\details{
Words are detected using Unicode-aware word boundaries as implemented in
\pkg{stringi}. This ensures robust handling of different languages,
punctuation, and whitespace.
}
\examples{
strCountWords("This is a sentence.")

strCountWords(c("One word", "Two words here", NA))

}
\seealso{
\code{\link[stringi:stri_count_words]{stringi::stri_count_words()}}

\link{string-overview} for an overview of all string utilities in pharos.
}
\concept{string-inspection}
\concept{summary}
