% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strLeftRight.R
\name{strLeftRight}
\alias{strLeftRight}
\alias{strLeft}
\alias{strRight}
\title{Returns the Left Or the Right Part Of a String}
\usage{
strLeft(x, n)

strRight(x, n)
}
\arguments{
\item{x}{a vector of strings.}

\item{n}{a positive or a negative integer, the number of characters to cut.
If n is negative, this number of characters will be cut off from the right
with \code{strLeft} and from the right with \code{strRight}. \cr n will be
recycled.}
}
\value{
the left (right) n characters of x
}
\description{
Returns the left part or the right part of a string. The number of
characters are defined by the argument \code{n}. If \code{n} is negative,
this number of characters will be cut off from the other side.
}
\details{
The functions \code{strLeft} and \code{strRight} are simple wrappers to
\code{substr}.
}
\examples{

strLeft("Hello world!", n=5)
strLeft("Hello world!", n=-5)

strRight("Hello world!", n=6)
strRight("Hello world!", n=-6)

strLeft(c("Lorem", "ipsum", "dolor","sit","amet"), n=2)

strRight(c("Lorem", "ipsum", "dolor","sit","amet"), n=c(2,3))

}
\seealso{
\code{\link{substr}}, \code{\link{strTrim}}

Other topic.text: 
\code{\link{strAbbr}()},
\code{\link{strAlign}()},
\code{\link{strCap}()},
\code{\link{strChop}()},
\code{\link{strCountW}()},
\code{\link{strDist}()},
\code{\link{strExtract}()},
\code{\link{strExtractBetween}()},
\code{\link{strIsNumeric}()},
\code{\link{strPad}()},
\code{\link{strPos}()},
\code{\link{strRev}()},
\code{\link{strSpell}()},
\code{\link{strSplit}()},
\code{\link{strTrim}()},
\code{\link{strTrunc}()},
\code{\link{strVal}()}
}
\concept{formatting}
\concept{string-processing}
\concept{topic.text}
