% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRev.R
\name{strRev}
\alias{strRev}
\title{Reverse Strings}
\usage{
strRev(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
A character vector with each string reversed.
}
\description{
Reverses each element of a character vector.
}
\details{
This function uses \code{\link[stringi:stri_reverse]{stringi::stri_reverse()}}, which correctly
handles Unicode characters and multi-byte encodings.
}
\examples{
strRev("abc")

strRev(c("hello", "world"))

# Unicode-safe
strRev("äöü")

}
\seealso{
\code{\link[stringi:stri_reverse]{stringi::stri_reverse()}}

\link{string-overview} for an overview of all string utilities in pharos.
}
\concept{string-manipulation}
