% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strSplit.R
\name{strSplit}
\alias{strSplit}
\title{Split Strings}
\usage{
strSplit(x, split = "", fixed = FALSE)
}
\arguments{
\item{x}{a character vector to be split}

\item{split}{a character string specifying the delimiter (if
\code{fixed = TRUE}) or a regular expression (if \code{fixed = FALSE})}

\item{fixed}{logical; if \code{TRUE}, \code{split} is treated as a fixed
string. Otherwise, it is interpreted as a regular expression.}
}
\value{
A list of character vectors. If \code{x} has length 1, a character
vector is returned.
}
\description{
Splits character vectors into substrings. This is a convenience wrapper
around \code{\link[stringi:stri_split_fixed]{stringi::stri_split_fixed()}} and
\code{\link[stringi:stri_split_regex]{stringi::stri_split_regex()}} with simplified defaults.
}
\details{
If the input \code{x} has length 1, the result is returned as a character
vector instead of a list for convenience.

This function provides a simplified interface to string splitting using
the \pkg{stringi} package. It avoids some of the complexity of
\code{\link[base:strsplit]{base::strsplit()}} while providing consistent and Unicode-aware
behavior.
}
\examples{
strSplit("a,b,c", ",", fixed = TRUE)

strSplit(c("a b", "c d"), " ")

# regex splitting
strSplit("a1b2c3", "\\\\d")

}
\seealso{
\code{\link[stringi:stri_split_fixed]{stringi::stri_split_fixed()}},
\code{\link[stringi:stri_split_regex]{stringi::stri_split_regex()}}, \code{\link[base:strsplit]{base::strsplit()}}

\link{string-overview} for an overview of all string utilities in pharos.
}
\concept{pattern-matching}
\concept{reshape}
\concept{string-manipulation}
