% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strTrim.R
\name{strTrim}
\alias{strTrim}
\title{Remove Leading/Trailing Whitespace From A String}
\usage{
strTrim(x, pattern = " \\t\\n", method = c("both", "left", "right"))
}
\arguments{
\item{x}{the string to be trimmed}

\item{pattern}{the pattern of the whitespaces to be deleted, defaults to
space, tab and newline: \code{" \\t\\n"}}

\item{method}{one out of \code{"both"} (default), \code{"left"},
\code{"right"}. Determines on which side the string should be trimmed.}
}
\value{
the string x without whitespaces
}
\description{
The function removes whitespace characters as spaces, tabs and newlines from
the beginning and end of the supplied string. Whitespace characters
occurring in the middle of the string are retained.\cr Trimming with method
\code{"left"} deletes only leading whitespaces, \code{"right"} only
trailing. Designed for users who were socialized by SQL.
}
\details{
The functions are defined depending on method as\cr \code{both: gsub( pattern=gettextf("^[\\\%s]+|[\\\%s]+$", pattern, pattern), replacement="", x=x)}\cr \code{left: gsub( pattern=gettextf("^[\\\%s]+",pattern), replacement="", x=x)}\cr \code{right: gsub( pattern=gettextf("[\\\%s]+$",pattern), replacement="", x=x)}
}
\examples{

strTrim("  Hello world! ")

strTrim("  Hello world! ", method="left")
strTrim("  Hello world! ", method="right")

# user defined pattern
strTrim(" ..Hello ... world! ", pattern=" \\\\.")

}
\seealso{
\code{\link[base:trimws]{base::trimws()}}

\link{string-overview} for an overview of all string utilities in pharos.
}
\concept{formatting}
\concept{string-manipulation}
