% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titleRect.R
\name{titleRect}
\alias{titleRect}
\title{Plot Boxed Annotation}
\usage{
titleRect(
  label,
  bg = "grey",
  border = 1,
  col = "black",
  xjust = 0.5,
  line = 2,
  ...
)
}
\arguments{
\item{label}{the main title}

\item{bg}{the background color of the box.}

\item{border}{the border color of the box}

\item{col}{the font color of the title}

\item{xjust}{the x-justification of the text. This can be \code{c(0, 0.5, 1)} for left, middle- and right alignement.}

\item{line}{on which MARgin line, starting at 0 counting outwards}

\item{\dots}{the dots are passed to the \code{\link[=text]{text()}} function, which
can be used to change font and similar arguments.}
}
\value{
nothing is returned
}
\description{
The function can be used to add a title to a plot surrounded by a
rectangular box. This is useful for plotting several plots in narrow
distances.
}
\examples{

plot(pressure)
titleRect("pressure")


}
\seealso{
\code{\link[=title]{title()}}

Other graphics.annotation:
\code{\link[=barText]{barText()}},
\code{\link[=boxedText]{boxedText()}},
\code{\link[=colLegend]{colLegend()}},
\code{\link[=errBars]{errBars()}},
\code{\link[=stamp]{stamp()}},
\code{\link[=textLegend]{textLegend()}}
}
\concept{annotation}
\concept{graphics.annotation}
\concept{label}
