% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformXY.R
\name{transformXY}
\alias{transformXY}
\title{Apply Geometric Transformations to Coordinates}
\usage{
transformXY(
  x,
  y = NULL,
  translate = c(0, 0),
  scale = c(1, 1),
  theta = 0,
  asp = 1
)
}
\arguments{
\item{x}{numeric vector of x coordinates, or an object coercible by
\code{\link[=xy.coords]{xy.coords()}}.}

\item{y}{numeric vector of y coordinates. Ignored if \code{x} already
contains both coordinates.}

\item{translate}{numeric vector of length 1 or 2 specifying translation
in x and y direction. Recycled if necessary. Default is \code{c(0, 0)}.}

\item{scale}{numeric vector of length 1 or 2 specifying scaling factors
for x and y. Recycled if necessary. Default is \code{c(1, 1)}.}

\item{theta}{rotation angle in radians. Default is \code{0}.}

\item{asp}{aspect ratio adjustment passed to \code{\link[=rotate]{rotate()}}.
Default is \code{1}.}
}
\value{
A list with components \code{x} and \code{y}, as returned by
\code{\link[=xy.coords]{xy.coords()}}.
}
\description{
Applies scaling, rotation, and translation to a set of 2D coordinates.
The transformations are applied in the following order:
\enumerate{
\item Scaling
\item Rotation (see \code{\link[=rotate]{rotate()}})
\item Translation
}
}
\details{
This function is a convenience wrapper combining basic affine transformations.
Internally, it uses \code{\link[=rotate]{rotate()}} for rotation.
}
\examples{
x <- c(0, 1, 1, 0)
y <- c(0, 0, 1, 1)

# scale, rotate and translate a square
transformXY(x, y,
            scale = 2,
            theta = pi / 4,
            translate = c(1, 1))

# matrix input
m <- cbind(x, y)
transformXY(m, scale = 0.5, theta = pi/6)


}
\seealso{
Other geometry.transformation:
\code{\link[=rotate]{rotate()}}
}
\concept{geometry}
\concept{geometry.transformation}
\concept{numerical-methods}
