% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcCoords.R
\name{abcCoords}
\alias{abcCoords}
\title{Coordinates for "bottomright", etc.}
\usage{
abcCoords(x = "topleft", region = "figure", cex = NULL, linset = 0, ...)
}
\arguments{
\item{x}{one out of \code{"bottomright"}, \code{"bottom"},
\code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"},
\code{"topright"}, \code{"right"}, \code{"center"}}

\item{region}{one out of \code{plot} or \code{figure}}

\item{cex}{the character extension for the text.}

\item{linset}{line inset in lines of text.}

\item{\dots}{the dots are passed to the \code{strwidth()} and
\code{strheight()} functions in case there where more specific text formats.}
}
\value{
nothing returned \%\% ~Describe the value returned \%\% If it is a LIST,
use \%\% \item{comp1 }{Description of 'comp1'}
}
\description{
Return the xy.coordinates for the literal positions "bottomright", etc. as
used to place legends.
}
\details{
The same logic as for the legend can be useful for placing texts, too. This
function returns the coordinates for the text, which can be used in the
specific text functions.
}
\examples{

plot(x = rnorm(10), type="n", xlab="", ylab="")
# note that plot.new() has to be called before we can grab the geometry
abcCoords("bottomleft")

lapply(c("bottomleft", "left"), abcCoords)

plot(x = rnorm(10), type="n", xlab="", ylab="")
text(x=(xy <- abcCoords("bottomleft", region = "plot"))$xy, 
     labels = "My Maybe Long Text", adj = xy$adj, xpd=NA)

text(x=(xy <- abcCoords("topleft", region = "figure"))$xy, 
     labels = "My Maybe Long Text", adj = xy$adj, xpd=NA)

plot(x = rnorm(10), type="n", xlab="", ylab="")
sapply(c("topleft", "top", "topright", "left", "center", 
         "right", "bottomleft", "bottom", "bottomright"), 
       function(x) 
         text(x=(xy <- abcCoords(x, region = "plot", linset=1))$xy, 
              labels = "MyMarginText", adj = xy$adj, xpd=NA)
)


plot(x = rnorm(100), type="n", xlab="", ylab="",
     panel.first={setBackCol(c("red", "lightyellow"))
             grid()})
xy <- abcCoords("topleft", region = "plot")
par(xpd=NA)
boxedText(x=xy$xy$x, y=xy$xy$y, xpad = 1, ypad = 1,
          labels = "My Maybe Long Text", adj = xy$adj, col=alpha("green", 0.8))

}
\seealso{
\code{\link{text}}, \code{\link{boxedText}}

Other plot.utils: 
\code{\link{axTicks}()},
\code{\link{lineToUser}()},
\code{\link{lines.lm}()},
\code{\link{mar}()},
\code{\link{setBackCol}()},
\code{\link{splineCI}},
\code{\link{spreadOut}()}
}
\concept{geometry}
\concept{graphics}
\concept{plot.utils}
