% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOpacity.R
\name{addOpacity}
\alias{addOpacity}
\title{Add an Alpha Channel to Colors}
\usage{
addOpacity(col, opacity = 0.5)
}
\arguments{
\item{col}{vector of valid R colors}

\item{opacity}{opacity value between 0 and 1}
}
\value{
Character vector of hexadecimal colors with alpha channel.
}
\description{
Add transparency to colors.
}
\examples{

op <- par(no.readonly = TRUE)

addOpacity("yellow", 0.2)
addOpacity(2, 0.5)   # red

canvas(3)
polygon(circle(x=c(-1,0,1), y=c(1,-1,1), radius=2), 
        col=addOpacity(2:4, 0.4))

x <- rnorm(15000)
par(mfrow=c(1,2))
plot(x, type="p", col="blue" )
plot(x, type="p", col=addOpacity("blue", .2), 
     main="Better insight with alpha channel" )

par(op)

}
\seealso{
\link[grDevices:adjustcolor]{grDevices::adjustcolor}

Other color.manipulation:
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=darken]{darken()}},
\code{\link[=fade]{fade()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}}
}
\concept{color}
\concept{color.manipulation}
\concept{transformation}
