% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{arc}
\alias{arc}
\title{Arc Geometry}
\usage{
arc(
  x = 0,
  y = 0,
  radiusX = 1,
  radiusY = radiusX,
  startAngle = 0,
  endAngle = 2 * pi,
  nPoints = 100
)
}
\arguments{
\item{x, y}{coordinates of the arc centre}

\item{radiusX, radiusY}{horizontal and vertical radius}

\item{startAngle, endAngle}{start and end angle in radians}

\item{nPoints}{number of points used to approximate the arc}
}
\value{
An object inheriting from class \code{"arcGeometry"}.
}
\description{
Create one or more circular or elliptic arcs.
}
\seealso{
Other geometry.structures:
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=circle]{circle()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=polygonX]{polygonX()}},
\code{\link[=regPolygon]{regPolygon()}},
\code{\link[=ring]{ring()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
