% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{arc}
\alias{arc}
\title{Arc Geometry}
\usage{
arc(
  x = 0,
  y = 0,
  radiusX = 1,
  radiusY = radiusX,
  startAngle = 0,
  endAngle = 2 * pi,
  numPoints = 100
)
}
\arguments{
\item{x, y}{Coordinates of the arc centre.}

\item{radiusX, radiusY}{Horizontal and vertical radius.}

\item{startAngle, endAngle}{Start and end angle in radians.}

\item{numPoints}{Number of points used to approximate the arc.}
}
\value{
An object inheriting from class \code{"arcGeometry"}.
}
\description{
Create one or more circular or elliptic arcs.
}
