% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.html.R
\name{as.html}
\alias{as.html}
\title{Mark a character vector as HTML}
\usage{
as.html(x)
}
\arguments{
\item{x}{a character vector, typically containing HTML markup}
}
\value{
\code{x}, with class \code{"html"} added
}
\description{
Tags a character vector with the S3 class \code{"html"} so that it
prints via \code{\link{preview.html}} as readable text instead of as
a raw character vector.
}
\examples{
as.html("<b>bold</b>")

}
\seealso{
\code{\link{preview.html}}, \code{\link{toHtmlTable}}
}
