% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisBreak.R
\name{axisBreak}
\alias{axisBreak}
\title{Place a Break Mark on an Axis}
\usage{
axisBreak(
  axis = 1,
  breakpos = NULL,
  pos = NA,
  bgcol = "white",
  breakcol = "black",
  style = "slash",
  brw = 0.02
)
}
\arguments{
\item{axis}{which axis to break.}

\item{breakpos}{where to place the break in user units.}

\item{pos}{position of the axis (see \link{axis}).}

\item{bgcol}{the color of the plot background.}

\item{breakcol}{the color of the "break" marker.}

\item{style}{Either \samp{gap}, \samp{slash} or \samp{zigzag}.}

\item{brw}{break width relative to plot width.}
}
\description{
Places a break mark on an axis on an existing plot.
}
\details{
The \samp{pos} argument is not needed unless the user has specified a
different position from the default for the axis to be broken.
}
\note{
There is some controversy about the propriety of using discontinuous
coordinates for plotting, and thus axis breaks. Discontinuous coordinates
allow widely separated groups of values or outliers to appear without
devoting too much of the plot to empty space. \cr The major objection seems
to be that the reader will be misled by assuming continuous coordinates.
The \samp{gap} style that clearly separates the two sections of the plot is
probably best for avoiding this.
}
\examples{

plot(3:10, main="Axis break test")

# put a break at the default axis and position
axisBreak()
axisBreak(2, 2.9, style="zigzag")

}
\seealso{
Other topic.graphics: 
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{canvas}()},
\code{\link{colorLegend}()},
\code{\link{errBars}()},
\code{\link{plot.Lc}()},
\code{\link{plotArea}()},
\code{\link{plotBag}()},
\code{\link{plotBar}()},
\code{\link{plotBubble}()},
\code{\link{plotCirc}()},
\code{\link{plotCor}()},
\code{\link{plotDens}()},
\code{\link{plotDens2D}()},
\code{\link{plotDot}()},
\code{\link{plotECDF}()},
\code{\link{plotFdist}()},
\code{\link{plotFun}()},
\code{\link{plotLines}()},
\code{\link{plotMiss}()},
\code{\link{plotMosaic}()},
\code{\link{plotProbDist}()},
\code{\link{plotQQ}()},
\code{\link{plotRidge}()},
\code{\link{plotTernary}()},
\code{\link{plotTimeSeries}()},
\code{\link{plotTreemap}()},
\code{\link{plotViolin}()},
\code{\link{plotWeb}()}
}
\author{
Jim Lemon and Ben Bolker
}
\concept{axis}
\concept{base-graphics}
\concept{plot-annotation}
\concept{topic.graphics}
