% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axTicks.R
\name{axTicks}
\alias{axTicks}
\alias{axTicks.POSIXct}
\alias{axTicks.Date}
\title{Compute Axis Tickmark Locations (For POSIXct Axis)}
\usage{
axTicks.POSIXct(side, x, at, format, labels = TRUE, ...)

axTicks.Date(side = 1, x, ...)
}
\arguments{
\item{side}{see \link[graphics:axis]{graphics::axis}}

\item{x, at}{date-time or date object.}

\item{format}{see \link[base:strptime]{base::strptime}}

\item{labels}{either a logical value specifying whether annotations are to
be made at the tickmarks, or a vector of character strings to be placed at
the tickpoints}

\item{\dots}{further arguments to be passed from or to other methods}
}
\value{
numeric vector of coordinate values at which axis tickmarks can be
drawn.
}
\description{
Compute pretty tickmark locations, the same way as R does internally. By
default, gives the at values which axis.POSIXct(side, x) would use.
}
\details{
\link{axTicks} has no implementation for POSIXct axis. This function fills
the gap.
}
\examples{

with(beaver1, {
  time <- strptime(paste(1990, day, time \%/\% 100, time \%\% 100),
                   "\%Y \%j \%H \%M")
  plot(time, temp, type = "l") # axis at 4-hour intervals.
  # now label every hour on the time axis
  plot(time, temp, type = "l", xaxt = "n")
  r <- as.POSIXct(round(range(time), "hours"))
  axis.POSIXct(1, at = seq(r[1], r[2], by = "hour"), format = "\%H")
  # place the grid
  abline(v=axTicks.POSIXct(1, at = seq(r[1], r[2], by = "hour"), format = "\%H"),
         col="grey", lty="dotted")
})

}
\seealso{
\code{\link[=axTicks]{axTicks()}}, \code{\link[=axis.POSIXct]{axis.POSIXct()}}

Other graphics.layout:
\code{\link[=abcCoords]{abcCoords()}},
\code{\link[=axisBreak]{axisBreak()}},
\code{\link[=isValidPlotRegion]{isValidPlotRegion()}},
\code{\link[=lineToUser]{lineToUser()}},
\code{\link[=mar]{mar()}},
\code{\link[=plotFacet]{plotFacet()}},
\code{\link[=spreadOut]{spreadOut()}}
}
\concept{date-time}
\concept{graphics.layout}
\concept{label}
