% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/band.R
\name{band}
\alias{band}
\title{Band Geometry}
\usage{
band(x, y)
}
\arguments{
\item{x}{A vector or matrix of x coordinates.}

\item{y}{A vector or matrix of y coordinates.

If either \code{x} or \code{y} is supplied as a two-column matrix,
the second column is interpreted as the lower boundary and reversed
automatically.}
}
\value{
An object inheriting from class
\code{"bandGeometry"}.
}
\description{
Create a polygonal band from upper and lower boundaries.
}
\details{
Typically used to represent confidence or prediction bands.
}
\examples{
set.seed(18)

x <- rnorm(15)
y <- x + rnorm(15)

new <- seq(-3, 3, 0.5)

pred <- predict(
  lm(y ~ x),
  newdata = data.frame(x = new),
  interval = "confidence"
)

plot(y ~ x)

polygon(
  band(
    x = new,
    y = pred[,2:3]
  ),
  col = addOpacity("grey80", 0.5),
  border = NA
)



}
\seealso{
\link[graphics:polygon]{graphics::polygon}

Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=circle]{circle()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=polygon]{polygon()}},
\code{\link[=regPolygon]{regPolygon()}},
\code{\link[=ring]{ring()}}
}
\concept{geometry}
\concept{geometry.structures}
