% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryTree.R
\name{plotBinaryTree}
\alias{plotBinaryTree}
\alias{binaryTree}
\title{Binary Tree}
\usage{
plotBinaryTree(
  x,
  main = "Binary tree",
  horiz = FALSE,
  text = TRUE,
  line = TRUE,
  ...
)
}
\arguments{
\item{x}{numeric vector to be organized as binary tree.}

\item{main}{main title of the plot.}

\item{horiz}{logical, should the plot be oriented horizontally or
vertically (default).}

\item{text}{properties of the text, can be any of the arguments
of \code{\link{boxedText}} (besides geometry and label).}

\item{line}{properties of the line segments (\code{col},
\code{lwd}, \code{lty}).}

\item{\dots}{the dots are sent to \code{\link{canvas}}.}
}
\value{
an integer vector of length n
}
\description{
Create a binary tree of a given number of nodes \code{n}. Can be used to
organize a sorted numeric vector as a binary tree.
}
\details{
If we index the nodes of the tree as 1 for the top, 2--3 for the next
horizontal row, 4--7 for the next, \ldots then the parent-child traversal
becomes particularly easy. The basic idea is that the rows of the tree start
at indices 1, 2, 4, \ldots.

binaryTree(13) yields the vector \code{c(8, 4, 9, 2, 10, 5, 11, 1, 12, 6, 13,
3, 7)} meaning that the smallest element will be in position 8 of the tree,
the next smallest in position 4, etc.
}
\note{
Substantially based on code by Terry Therneau, with major extensions
and improvements by the package author.
}
\examples{

bedrock::binaryTree(12)

x <- sort(sample(100, 24))
z <- plotBinaryTree(x, cex=0.8)

plotBinaryTree(LETTERS[1:15], 
               text=list(col="deeppink4", cex=2),
               line=list(col="navajowhite3", lwd=2))


# Plot example - Titanic data, for once from a somwhat different perspective
tab <- apply(Titanic, c(2,3,4), sum)
cprob <- c(1, prop.table(apply(tab, 1, sum))
           , as.vector(aperm(prop.table(apply(tab, c(1,2), sum), 1), c(2, 1)))
           , as.vector(aperm(prop.table(tab, c(1,2)), c(3,2,1)))
)

plotBinaryTree(round(cprob[bedrock::binaryTree(length(cprob))],2), horiz=TRUE, cex=0.8,
            main="Titanic")
text(c("sex","age","survived"), y=0, x=c(1,2,3)+1)

}
\seealso{
Other plot.special:
\code{\link[=plotArea]{plotArea()}},
\code{\link[=plotCirc]{plotCirc()}},
\code{\link[=plotFun]{plotFun()}},
\code{\link[=plotMiss]{plotMiss()}},
\code{\link[=plotPolar]{plotPolar()}},
\code{\link[=plotTernary]{plotTernary()}},
\code{\link[=plotTreemap]{plotTreemap()}},
\code{\link[=plotWeb]{plotWeb()}}
}
\concept{data-structures}
\concept{graphics}
\concept{plot.special}
