% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxedText.R
\name{boxedText}
\alias{boxedText}
\alias{boxedText.default}
\alias{boxedText.formula}
\title{Add Text in a Box to a Plot}
\usage{
boxedText(x, ...)

\method{boxedText}{default}(
  x,
  y = NULL,
  labels = NULL,
  adj = NULL,
  pos = NULL,
  offset = 0.5,
  vfont = NULL,
  cex = 1,
  col = NULL,
  font = NULL,
  srt = 0,
  xPad = 0.2,
  yPad = 0.2,
  density = NULL,
  angle = 45,
  bg = NA,
  border = par("fg"),
  lty = par("lty"),
  lwd = par("lwd"),
  ...
)

\method{boxedText}{formula}(formula, data = parent.frame(), ..., subset)
}
\arguments{
\item{x, y}{numeric vectors of coordinates where the text labels should be
written. If the length of x and y differs, the shorter one is recycled.}

\item{\dots}{additional arguments are passed to the text function.}

\item{labels}{a character vector or expression specifying the text to be
written.  An attempt is made to coerce other language objects (names and
calls) to expressions, and vectors and other classed objects to character
vectors by as.character. If labels is longer than x and y, the coordinates
are recycled to the length of labels.}

\item{adj}{the value of adj determines the way in which text strings are
justified.  A value of 0 produces left-justified text, 0.5 (the default)
centered text and 1 right-justified text.  (Any value in \verb{[0, 1]} is allowed,
and on most devices values outside that interval will also work.)  Note that
the adj argument of text also allows adj = c(x, y) for different adjustment
in x- and y- directions.}

\item{pos}{a position specifier for the text. If specified this overrides
any adj value given. Values of 1, 2, 3 and 4, respectively indicate
positions below, to the left of, above and to the right of the specified
coordinates.}

\item{offset}{when pos is specified, this value gives the offset of the
label from the specified coordinate in fractions of a character width.}

\item{vfont}{\code{NULL} for the current font family, or a character vector
of length 2 for Hershey vector fonts. The first element of the vector
selects a typeface and the second element selects a style. Ignored if labels
is an expression.}

\item{cex}{numeric character expansion factor; multiplied by
\code{par("cex")} yields the final character size. \code{NULL} and \code{NA}
are equivalent to 1.0.}

\item{col, font}{the color and (if vfont = NULL) font to be used, possibly
vectors. These default to the values of the global graphical parameters in
\code{par()}.}

\item{srt}{the string rotation in degrees.}

\item{xPad, yPad}{the proportion of the rectangles to the extent of the text
within.}

\item{density}{the density of shading lines, in lines per inch. The default
value of \code{NULL} means that no shading lines are drawn.  A zero value of
density means no shading lines whereas negative values (and NA) suppress
shading (and so allow color filling).}

\item{angle}{angle (in degrees) of the shading lines.}

\item{bg}{color(s) to fill or shade the rectangle(s) with. The default
\code{NA} (or also NULL) means do not fill, i.e., draw transparent
rectangles, unless density is specified.}

\item{border}{color for rectangle border(s). The default is
\code{par("fg")}. Use \code{border = NA} to omit borders (this is the
default).  If there are shading lines, \code{border = TRUE} means use the
same colour for the border as for the shading lines.}

\item{lty}{line type for borders and shading; defaults to \code{"solid"}.}

\item{lwd}{line width for borders and shading. Note that the use of
\code{lwd = 0} (as in the examples) is device-dependent.}

\item{formula}{A formula of the form \code{lhs ~ rhs}, where \code{lhs}
gives the response values and \code{rhs} the corresponding groups
or explanatory variables.}

\item{data}{an optional matrix or data frame (or similar; see
\code{\link[stats:model.frame]{stats::model.frame()}}) containing the variables in the
formula. By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{an optional vector specifying a subset of observations
to be used in the analysis.}
}
\value{
No return value, called for its side effect of adding the boxed
labels to the current plot.
}
\description{
boxedText draws the strings given in the vector labels at the coordinates
given by x and y, surrounded by a rectangle.
}
\examples{

canvas(xpd=TRUE)

boxedText(0, 0, adj=0, label="This is boxed text", srt=seq(0,360,20), 
          xPad=.3, yPad=.3)
points(0,0, pch=15)

plot(mpg ~ hp, data=mtcars, type="n", main="MT cars mpg/hp (log-log)", 
     panel.first=quote(grid()), las=1, log="xy")
boxedText(mpg ~ hp, data=mtcars, 
          labels=rownames(mtcars), cex=0.6, border=FALSE, bg="grey90")


}
\seealso{
similar function in package \pkg{plotrix}
\code{plotrix::boxed.labels} (lacking rotation option)

Other graphics.annotation:
\code{\link[=barText]{barText()}},
\code{\link[=colLegend]{colLegend()}},
\code{\link[=errBars]{errBars()}},
\code{\link[=stamp]{stamp()}},
\code{\link[=textLegend]{textLegend()}},
\code{\link[=titleRect]{titleRect()}}
}
\concept{annotation}
\concept{graphics.annotation}
\concept{label}
