% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\title{Canvas for Geometric Plotting}
\usage{
canvas(xlim = NULL, ylim = xlim, main = NULL, asp = 1, usrbg = "white", ...)
}
\arguments{
\item{xlim, ylim}{the xlims and ylims for the plot. Default is c(-1, 1).}

\item{main}{the main title on top of the plot.}

\item{asp}{numeric, giving the aspect ratio y/x. (See
\code{\link{plot.window}} for details. Default is 1.}

\item{usrbg}{the color of the user space of the plot, defaults to "white".}

\item{\dots}{additional arguments are passed to the \code{plot()} command.}
}
\value{
a list of all the previous values of the parameters changed
(returned invisibly)
}
\description{
This is just a wrapper for creating an empty plot with suitable defaults for
plotting geometric shapes.
}
\details{
The plot is created with these settings:\cr \code{asp = 1, xaxt = "n", yaxt
= "n", xlab = "", ylab = "", frame.plot = FALSE}.
}
\examples{

canvas(7)
text(0, 0, "Hello world!", cex=5)

}
\seealso{
Other topic.graphics: 
\code{\link{axisBreak}()},
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{colorLegend}()},
\code{\link{errBars}()},
\code{\link{plot.Desc.factor}()},
\code{\link{plot.Desc.logical}()},
\code{\link{plot.Desc.table}()},
\code{\link{plot.Desc.ts}()},
\code{\link{plot.Lc}()},
\code{\link{plot.PostHocTest}()},
\code{\link{plotArea}()},
\code{\link{plotBag}()},
\code{\link{plotBar}()},
\code{\link{plotBubble}()},
\code{\link{plotCirc}()},
\code{\link{plotCor}()},
\code{\link{plotDens}()},
\code{\link{plotDens2D}()},
\code{\link{plotDot}()},
\code{\link{plotECDF}()},
\code{\link{plotFdist}()},
\code{\link{plotFun}()},
\code{\link{plotLines}()},
\code{\link{plotMiss}()},
\code{\link{plotMosaic}()},
\code{\link{plotProbDist}()},
\code{\link{plotQQ}()},
\code{\link{plotRidge}()},
\code{\link{plotTernary}()},
\code{\link{plotTimeSeries}()},
\code{\link{plotTreemap}()},
\code{\link{plotViolin}()},
\code{\link{plotWeb}()}
}
\concept{base-graphics}
\concept{layout}
\concept{topic.graphics}
