% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{circle}
\alias{circle}
\title{Circle Geometry}
\usage{
circle(x = 0, y = 0, radius = 1, nPoints = 100)
}
\arguments{
\item{x, y}{centre coordinates.}

\item{radius}{circle radius.}

\item{nPoints}{number of points used to approximate the circle.}
}
\value{
An object inheriting from class \code{"circleGeometry"}.
}
\description{
Create a circular geometry.
}
\examples{
canvas(xlim = c(-2, 2))

polygonX(circle(radius = 1.5), col = "lightblue")

# several circles at once
polygonX(circle(x = c(-0.6, 0.6), y = 0, radius = 0.4),
         col = c("tomato", "royalblue"))

}
\seealso{
Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=polygonX]{polygonX()}},
\code{\link[=regPolygon]{regPolygon()}},
\code{\link[=ring]{ring()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
