% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{circle}
\alias{circle}
\title{Circle Geometry}
\usage{
circle(x = 0, y = 0, radius = 1, numPoints = 100)
}
\arguments{
\item{x, y}{Centre coordinates.}

\item{radius}{Circle radius.}

\item{numPoints}{Number of points used to approximate the circle.}
}
\value{
An object inheriting from class \code{"circleGeometry"}.
}
\description{
Create a circular geometry.
}
