% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{CMYKToCMY}
\alias{CMYKToCMY}
\title{Convert CMYK to CMY}
\usage{
CMYKToCMY(col)
}
\arguments{
\item{col}{Numeric CMYK matrix.}
}
\value{
Numeric CMY matrix.
}
\description{
Convert CMYK colors to CMY.
}
\seealso{
Other color.space:
\code{\link[=CMYKToRGB]{CMYKToRGB()}},
\code{\link[=CMYToCMYK]{CMYToCMYK()}},
\code{\link[=RGBToCMY]{RGBToCMY()}}
}
\concept{color-manipulation}
\concept{color.space}
