% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{cmykToCmy}
\alias{cmykToCmy}
\title{Convert CMYK to CMY}
\usage{
cmykToCmy(col)
}
\arguments{
\item{col}{Numeric CMYK matrix.}
}
\value{
Numeric CMY matrix.
}
\description{
Convert CMYK colors to CMY.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color-conversion}
\concept{color-space}
\concept{graphics}
