% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{CMYKToRGB}
\alias{CMYKToRGB}
\title{Convert CMYK to RGB}
\usage{
CMYKToRGB(cyan, magenta, yellow, black, maxColorValue = 1)
}
\arguments{
\item{cyan}{Cyan values.}

\item{magenta}{Magenta values.}

\item{yellow}{Yellow values.}

\item{black}{Black values.}

\item{maxColorValue}{Maximum channel value.}
}
\value{
Hexadecimal RGB colors.
}
\description{
Convert CMYK colors to RGB.
}
\seealso{
Other color.space:
\code{\link[=CMYKToCMY]{CMYKToCMY()}},
\code{\link[=CMYToCMYK]{CMYToCMYK()}},
\code{\link[=RGBToCMY]{RGBToCMY()}}
}
\concept{color-conversion}
\concept{color-space}
\concept{color.space}
\concept{graphics}
