% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{cmykToRgb}
\alias{cmykToRgb}
\title{Convert CMYK to RGB}
\usage{
cmykToRgb(cyan, magenta, yellow, black, maxColorValue = 1)
}
\arguments{
\item{cyan}{Cyan values.}

\item{magenta}{Magenta values.}

\item{yellow}{Yellow values.}

\item{black}{Black values.}

\item{maxColorValue}{Maximum channel value.}
}
\value{
Hexadecimal RGB colors.
}
\description{
Convert CMYK colors to RGB.
}
\seealso{
Other color.space:
\code{\link[=cmyToCmyk]{cmyToCmyk()}},
\code{\link[=cmykToCmy]{cmykToCmy()}},
\code{\link[=rgbToCmy]{rgbToCmy()}}
}
\concept{color-conversion}
\concept{color-space}
\concept{color.space}
\concept{graphics}
