% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{CMYToCMYK}
\alias{CMYToCMYK}
\title{Convert CMY to CMYK}
\usage{
CMYToCMYK(col)
}
\arguments{
\item{col}{Numeric CMY matrix.}
}
\value{
Numeric CMYK matrix.
}
\description{
Convert CMY colors to CMYK.
}
\seealso{
Other color.space:
\code{\link[=CMYKToCMY]{CMYKToCMY()}},
\code{\link[=CMYKToRGB]{CMYKToRGB()}},
\code{\link[=RGBToCMY]{RGBToCMY()}}
}
\concept{color-conversion}
\concept{color-space}
\concept{color.space}
\concept{graphics}
