% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{cmyToCmyk}
\alias{cmyToCmyk}
\title{Convert CMY to CMYK}
\usage{
cmyToCmyk(col)
}
\arguments{
\item{col}{Numeric CMY matrix.}
}
\value{
Numeric CMYK matrix.
}
\description{
Convert CMY colors to CMYK.
}
\seealso{
Other color.space:
\code{\link[=cmykToCmy]{cmykToCmy()}},
\code{\link[=cmykToRgb]{cmykToRgb()}},
\code{\link[=rgbToCmy]{rgbToCmy()}}
}
\concept{color-conversion}
\concept{color-space}
\concept{color.space}
\concept{graphics}
