% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{colToHex}
\alias{colToHex}
\title{Convert R Colors to Hexadecimal Colors}
\usage{
colToHex(col, opacity = 1)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{opacity}{Opacity value between 0 and 1.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Convert any valid R color specification to hexadecimal colors.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color-conversion}
\concept{graphics}
