% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{colToHex}
\alias{colToHex}
\title{Convert R Colors to Hexadecimal Colors}
\usage{
colToHex(col, alpha = 1)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{alpha}{Alpha transparency value.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Convert any valid R color specification to hexadecimal colors.
}
\seealso{
Other color.conversion:
\code{\link[=RGBToCol]{RGBToCol()}},
\code{\link[=RGBToHex]{RGBToHex()}},
\code{\link[=RGBToLong]{RGBToLong()}},
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=longToRGB]{longToRGB()}}
}
\concept{color-conversion}
\concept{color.conversion}
\concept{graphics}
