% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{colToHSV}
\alias{colToHSV}
\title{Convert R Colors to HSV}
\usage{
colToHSV(col, useAlphaChannel = FALSE)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{useAlphaChannel}{Logical indicating whether the alpha
channel should be included.}
}
\value{
Numeric HSV matrix.
}
\description{
Convert any valid R color specification to HSV.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color-conversion}
\concept{graphics}
