% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{colToHSV}
\alias{colToHSV}
\title{Convert R Colors to HSV}
\usage{
colToHSV(col, alpha = FALSE)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{alpha}{Logical indicating whether alpha values should
be returned.}
}
\value{
Numeric HSV matrix.
}
\description{
Convert any valid R color specification to HSV.
}
\seealso{
Other color.conversion:
\code{\link[=colToHex]{colToHex()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=longToRGB]{longToRGB()}},
\code{\link[=rgbToCol]{rgbToCol()}},
\code{\link[=rgbToHex]{rgbToHex()}},
\code{\link[=rgbToLong]{rgbToLong()}}
}
\concept{color-conversion}
\concept{color.conversion}
\concept{graphics}
