% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colToOpaque.R
\name{colToOpaque}
\alias{colToOpaque}
\title{Equivalent Opaque Color for Transparent Color}
\usage{
colToOpaque(col, alpha = NULL, bg = NULL)
}
\arguments{
\item{col}{the color as hex value (use converters below if it's not
available). \code{col} and \code{alpha} are recycled.}

\item{alpha}{the alpha channel, if left to NULL the alpha channels of the
colors are used}

\item{bg}{the background color to be used to calculate against (default is
"white")}
}
\value{
An named vector with the hexcodes of the opaque colors.
}
\description{
Determine the equivalent opaque RGB color for a given partially transparent
RGB color against a background of any color.
}
\details{
Reducing the opacity against a white background is a good way to find usable
lighter and less saturated tints of a base color. For doing so, we sometimes
need to get the equivalent opaque color for the transparent color.
}
\examples{

cols <- c(alpha("limegreen", 0.4), colToOpaque(colToHex("limegreen"), 0.4), "limegreen")
barplot(c(1, 1.2, 1.3), col=cols, panel.first=abline(h=0.4, lwd=10, col="grey35"))
}
\seealso{
\code{\link{colToHex}}, \code{\link[bedrock]{decToHex}},
\code{\link{rgbToHex}}

Other topic.colors: 
\code{\link{alpha}()},
\code{\link{contrastColor}()},
\code{\link{conv_cols}},
\code{\link{conv_rgb}},
\code{\link{findColor}()}
}
\concept{color-conversion}
\concept{color-manipulation}
\concept{topic.colors}
