% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{colToRGB}
\alias{colToRGB}
\title{Convert R Colors to RGB}
\usage{
colToRGB(col, alpha = FALSE)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{alpha}{Logical indicating whether the alpha channel
should be included.}
}
\value{
Integer matrix with rows corresponding to RGB
(and optionally alpha) channels.
}
\description{
Convert any valid R color specification to an RGB matrix.
}
\examples{

rgbToCol(matrix(c(162,42,42), nrow=3))

rgbToLong(matrix(c(162,42,42), nrow=3))

colToRGB("peachpuff")
colToRGB(c(blu = "royalblue", reddish = "tomato")) # names kept

colToRGB(1:8)



}
\seealso{
\code{\link{col2rgb}}

Other color:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=contrastColor]{contrastColor()}},
\code{\link[=darken]{darken()}},
\code{\link[=findColor]{findColor()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=hcol]{hcol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}},
\code{\link[=pal]{pal()}},
\code{\link[=palNames]{palNames()}},
\code{\link[=setBackCol]{setBackCol()}}
}
\concept{color}
\concept{color-conversion}
