% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{colToRGB}
\alias{colToRGB}
\title{Convert R Colors to RGB}
\usage{
colToRGB(col, useAlphaChannel = FALSE)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{useAlphaChannel}{Logical indicating whether the alpha
channel should be included.}
}
\value{
Integer matrix with rows corresponding to RGB
(and optionally alpha) channels.
}
\description{
Convert any valid R color specification to an RGB matrix.
}
\examples{

rgbToCol(matrix(c(162,42,42), nrow=3))

rgbToLong(matrix(c(162,42,42), nrow=3))

colToRGB("peachpuff")
colToRGB(c(blu = "royalblue", reddish = "tomato")) # names kept

colToRGB(1:8)


}
\seealso{
\link[grDevices:col2rgb]{grDevices::col2rgb}, \link{color-conversion-overview}
}
\concept{color}
