% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{colToRGB}
\alias{colToRGB}
\title{Convert R Colors to RGB}
\usage{
colToRGB(col, alpha = FALSE)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{alpha}{Logical indicating whether the alpha channel
should be included.}
}
\value{
Integer matrix with rows corresponding to RGB
(and optionally alpha) channels.
}
\description{
Convert any valid R color specification to an RGB matrix.
}
\examples{

RGBToCol(matrix(c(162,42,42), nrow=3))

RGBToLong(matrix(c(162,42,42), nrow=3))

colToRGB("peachpuff")
colToRGB(c(blu = "royalblue", reddish = "tomato")) # names kept

colToRGB(1:8)



}
\seealso{
\code{\link{col2rgb}}

Other color.conversion:
\code{\link[=RGBToCol]{RGBToCol()}},
\code{\link[=RGBToHex]{RGBToHex()}},
\code{\link[=RGBToLong]{RGBToLong()}},
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToHex]{colToHex()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=longToRGB]{longToRGB()}}
}
\concept{color-manipulation}
\concept{color.conversion}
