% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{Prefix}
\title{DescToolsX Constants}
\format{
A character vector of length 7.
}
\usage{
Prefix
}
\description{
English abbreviated and full weekday names, starting with Monday.
Metric prefixes with name, abbreviation and dimension.
}
\keyword{internal}
