% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darken.R
\name{darken}
\alias{darken}
\title{Darken Colors}
\usage{
darken(col, amount = 0.2)
}
\arguments{
\item{col}{vector of valid R colors.}

\item{amount}{numeric value between 0 and 1 specifying the
amount of darkening. A value of 0 leaves the color unchanged,
while 1 returns black.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Darken colors by mixing them with black.
}
\details{
Colors are mixed linearly with black in RGB space:
\deqn{
  x_{new} = x \cdot (1 - amount)
}
}
\examples{
darken("tomato")
darken(c("tomato", "royalblue"), amount = 0.5)

# increasing amounts
amount <- seq(0, 1, 0.2)
barplot(rep(1, 6), col = sapply(amount, darken, col = "tomato"),
        names.arg = amount, yaxt = "n", xlab = "amount")

}
\seealso{
Other color.manipulation:
\code{\link[=addOpacity]{addOpacity()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=fade]{fade()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixCol]{mixCol()}}
}
\concept{color}
\concept{color.manipulation}
\concept{transformation}
