% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darken.R
\name{darken}
\alias{darken}
\title{Darken Colors}
\usage{
darken(col, amount = 0.2)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{amount}{Numeric value between 0 and 1 specifying the
amount of darkening. A value of 0 leaves the color unchanged,
while 1 returns black.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Darken colors by mixing them with black.
}
\details{
Colors are mixed linearly with black in RGB space:
\deqn{
  x_{new} = x \\cdot (1 - amount)
}
}
\seealso{
Other color:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=contrastColor]{contrastColor()}},
\code{\link[=findColor]{findColor()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=hcol]{hcol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}},
\code{\link[=pal]{pal()}},
\code{\link[=palNames]{palNames()}},
\code{\link[=setBackCol]{setBackCol()}}
}
\concept{color}
