% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degToRad.R
\name{degToRad}
\alias{degToRad}
\alias{radToDeg}
\title{Convert Degrees to Radians and Vice Versa}
\usage{
degToRad(deg)

radToDeg(rad)
}
\arguments{
\item{deg}{a vector of angles in degrees.}

\item{rad}{a vector of angles in radians.}
}
\value{
degToRad returns a vector of the same length as \code{deg} with the
angles in radians.\cr radToDeg returns a vector of the same length as
\code{rad} with the angles in degrees.
}
\description{
Convert degrees to radians (and back again).
}
\examples{

degToRad(c(90,180,270))
radToDeg( c(0.5,1,2) * pi)

}
\seealso{
Other graphics.utils:
\code{\link[=abcCoords]{abcCoords()}},
\code{\link[=axisBreak]{axisBreak()}},
\code{\link[=barText]{barText()}},
\code{\link[=boxedText]{boxedText()}},
\code{\link[=colLegend]{colLegend()}},
\code{\link[=errBars]{errBars()}},
\code{\link[=lineSep]{lineSep()}},
\code{\link[=lineToUser]{lineToUser()}},
\code{\link[=lines.lm]{lines.lm()}},
\code{\link[=mar]{mar()}},
\code{\link[=preview]{preview()}},
\code{\link[=spreadOut]{spreadOut()}},
\code{\link[=stamp]{stamp()}},
\code{\link[=textLegend]{textLegend()}},
\code{\link[=titleRect]{titleRect()}},
\code{\link[=unit]{unit()}}
}
\concept{annotation}
\concept{graphics.utils}
