% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawBand.R
\name{drawBand}
\alias{drawBand}
\title{Draw Confidence Band}
\usage{
drawBand(x, y, col = alpha("grey", 0.5), border = NA)
}
\arguments{
\item{x}{a vector or a matrix with x coordinates for the band. If x is given
as matrix it must be a \eqn{2 \times n}{2 x n} matrix and the second column
will be reversed. x will be recyled in the case y is a 2dimensional matrix.}

\item{y}{a vector or a matrix with y coordinates for the band. If y is given
as matrix it must be a \eqn{2 \times n}{2 x n} matrix and the second column
will be reversed. y will be recyled in the case x is a 2dimensional matrix.}

\item{col}{the color of the band.}

\item{border}{the border color of the band.}
}
\description{
Draw a band using a simple syntax. Just a wrapper for the function
\code{polygon()} typically used to draw confidence bands.
}
\examples{

set.seed(18)
x <- rnorm(15)
y <- x + rnorm(15)

new <- seq(-3, 3, 0.5)
pred.w.plim <- predict(lm(y ~ x), newdata=data.frame(x=new), interval="prediction")
pred.w.clim <- predict(lm(y ~ x), newdata=data.frame(x=new), interval="confidence")

plot(y ~ x)
drawBand(y = c(pred.w.plim[,2], rev(pred.w.plim[,3])),
  x=c(new, rev(new)), col= alpha("grey90", 0.5))

# passing y as matrix interface allows more intuitive arguments
drawBand(y = pred.w.clim[, 2:3],
         x = new, col= alpha("grey80", 0.5))

abline(lm(y~x), col="brown")

}
\seealso{
\code{\link{polygon}}

Other plot.annotation:
\code{\link[=axisBreak]{axisBreak()}},
\code{\link[=barText]{barText()}},
\code{\link[=boxedText]{boxedText()}},
\code{\link[=colorLegend]{colorLegend()}},
\code{\link[=errBars]{errBars()}},
\code{\link[=shade]{shade()}},
\code{\link[=stamp]{stamp()}},
\code{\link[=titleRect]{titleRect()}}
}
\concept{confidence-intervals}
\concept{graphics}
\concept{plot.annotation}
