% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawBezier.R
\name{drawBezier}
\alias{drawBezier}
\title{Draw a Bezier Curve}
\usage{
drawBezier(
  x = 0,
  y = x,
  nv = 100,
  col = par("col"),
  lty = par("lty"),
  lwd = par("lwd"),
  plot = TRUE
)
}
\arguments{
\item{x, y}{a vector of xy-coordinates to define the Bezier curve. Should at
least contain 3 points.}

\item{nv}{number of vertices to draw the curve.}

\item{col}{color(s) for the curve. Default is \code{par("fg")}.}

\item{lty}{line type for borders and shading; defaults to \code{"solid"}.}

\item{lwd}{line width for borders and shading.}

\item{plot}{logical. If \code{TRUE} the structure will be plotted. If
\code{FALSE} only the xy-points are calculated and returned.  Use this if
you want to combine several geometric structures to a single polygon.}
}
\value{
\code{drawBezier} invisibly returns a list of the calculated
coordinates for all shapes.
}
\description{
Draw a Bezier curve.
}
\details{
Bezier curves appear in such areas as mechanical computer aided design
(CAD).  They are named after P. Bezier, who used a closely related
representation in Renault's UNISURF CAD system in the early 1960s (similar,
unpublished, work was done by P. de Casteljau at Citroen in the late 1950s
and early 1960s). The 1970s and 1980s saw a flowering of interest in Bezier
curves, with many CAD systems using them, and many important developments in
their theory.  The usefulness of Bezier curves resides in their many
geometric and analytical properties.  There are elegant and efficient
algorithms for evaluation, differentiation, subdivision of the curves, and
conversion to other useful representations. (See: Farin, 1993)
}
\examples{

canvas(xlim=c(0,1))
grid()
drawBezier( x=c(0,0.5,1), y=c(0,0.5,0), col="blue", lwd=2)
drawBezier( x=c(0,0.5,1), y=c(0,1,0), col="red", lwd=2)
drawBezier( x=c(0,0.25,0.5,0.75,1), y=c(0,1,1,1,0), col="darkgreen", lwd=2)
}
\references{
G. Farin (1993) \emph{Curves and surfaces for computer aided
geometric design. A practical guide}, Acad. Press
}
\seealso{
\code{\link{polygon}}, \code{\link{drawRegPolygon}},
\code{\link{drawCircle}}, \code{\link{drawArc}}

Other plot.geometry:
\code{\link[=canvas]{canvas()}},
\code{\link[=drawArc]{drawArc()}},
\code{\link[=drawCircle]{drawCircle()}},
\code{\link[=drawEllipse]{drawEllipse()}},
\code{\link[=drawRegPolygon]{drawRegPolygon()}},
\code{\link[=polarGrid]{polarGrid()}}
}
\author{
Frank E Harrell Jr \href{mailto:f.harrell@vanderbilt.edu}{f.harrell@vanderbilt.edu}
}
\concept{geometry}
\concept{graphics}
\concept{plot.geometry}
