% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCircle.R
\name{drawCircle}
\alias{drawCircle}
\title{Draw a Circle}
\usage{
drawCircle(
  x = 0,
  y = x,
  r.out = 1,
  r.in = 0,
  theta.1 = 0,
  theta.2 = 2 * pi,
  border = par("fg"),
  col = NA,
  lty = par("lty"),
  lwd = par("lwd"),
  nv = 100,
  plot = TRUE
)
}
\arguments{
\item{x, y}{a vector (or scalar) of xy-coordinates for the center(s) of the
circle(s).}

\item{r.out}{a vector (or scalar) of the outer radius of the circle.}

\item{r.in}{a vector (or scalar) of a potential inner radius of an annulus.}

\item{theta.1}{a vector (or scalar) of the starting angle(s). The sectors
are built counterclockwise.}

\item{theta.2}{a vector (or scalar) of the ending angle(s).}

\item{border}{color for circle borders. The default is par("fg"). Use border
= \code{NA} to omit borders.}

\item{col}{color(s) to fill or shade the circle(s) with. The default
\code{NA} (or also NULL) means do not fill, i.e., draw transparent circles,
unless density is specified.}

\item{lty}{line type for borders and shading; defaults to \code{"solid"}.}

\item{lwd}{line width for borders and shading.}

\item{nv}{number of vertices to draw the circle.}

\item{plot}{logical. If \code{TRUE} the structure will be plotted. If
\code{FALSE} only the points are calculated and returned. Use this option if
you want to combine several geometric structures to a polygon.}
}
\value{
The function invisibly returns a list of the calculated coordinates
for all shapes.
}
\description{
Draw one or several circle on an existing plot.
}
\details{
All geometric arguments will be recycled.
}
\examples{

canvas(xlim = c(-5,5), xpd=TRUE)
cols <- pal("Helsana")[1:5]

# Draw ring
drawCircle (r.in = 1, r.out = 5, border="darkgrey", 
            col=alpha("yellow", 0.2), lwd=2)

# Draw circle
drawCircle (r.in = 6, border="green", lwd=3)

# Draw sectors
geom <- rbind(c(-pi, 0, .25, .5), c(0, pi, 1, 2),
              c(-pi/2, pi/2, 2, 2.5), c(pi/2, 3 * pi/2, 3, 4),
              c(pi - pi/8, pi + pi/8, 1.5, 2.5))

drawCircle (r.in = geom[,3], r.out = geom[,4],
           theta.1 = geom[,1], theta.2 = geom[,2],
           col = alpha(cols, 0.6),
           border = cols, lwd=1)


# clipping
canvas(bg="lightgrey", main="Yin ~ Yang")
drawCircle (r.out = 1, col="white")
clip(0, 2, 2, -2)
drawCircle(col="black")
clip(-2, 2, 2, -2)
drawCircle (y = c(-0.5,0.5), r.out = 0.5, col=c("black", "white"), border=NA)
drawCircle (y = c(-0.5,0.5), r.out = 0.1, col=c("white", "black"), border=NA)
drawCircle ()


# overplotting circles
canvas(xlim=c(-5,5))
drawCircle (r.out=4:1, col=c("white", "steelblue2", "white", "red"), lwd=3, nv=300)


# rotation
x <- seq(-3, 3, length.out=10)
y <- rep(0, length.out=length(x))

canvas(xlim=c(-5,5), bg="black")

sapply( (0:11) * pi/6, function(theta) {
  xy <- rotate(x, y=y, theta=theta)
  drawCircle (x=xy$x, y=xy$y, r.in=2.4, border=alpha("white", 0.2))
} )
}
\seealso{
\code{\link{polygon}}, \code{\link{drawRegPolygon}},
\code{\link{drawEllipse}}, \code{\link{drawArc}}

Other plot.geometry: 
\code{\link{canvas}()},
\code{\link{drawArc}()},
\code{\link{drawBezier}()},
\code{\link{drawEllipse}()},
\code{\link{drawRegPolygon}()},
\code{\link{polarGrid}()}
}
\concept{geometry}
\concept{graphics}
\concept{plot.geometry}
