% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{ellipse}
\alias{ellipse}
\title{Ellipse Geometry}
\usage{
ellipse(x = 0, y = 0, radiusX = 1, radiusY = radiusX, nPoints = 100)
}
\arguments{
\item{x, y}{centre coordinates.}

\item{radiusX, radiusY}{horizontal and vertical radius.}

\item{nPoints}{number of points used to approximate the ellipse.}
}
\value{
An object inheriting from class \code{"ellipseGeometry"}.
}
\description{
Create an elliptic geometry.
}
\details{
Use \code{\link[=rotate]{rotate()}} to rotate the resulting geometry.
}
\seealso{
Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=circle]{circle()}},
\code{\link[=polygonX]{polygonX()}},
\code{\link[=regPolygon]{regPolygon()}},
\code{\link[=ring]{ring()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
