% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errBars.R
\name{errBars}
\alias{errBars}
\title{Add Error Bars to an Existing Plot}
\usage{
errBars(
  from,
  to = NULL,
  pos = NULL,
  mid = NULL,
  horiz = FALSE,
  col = par("fg"),
  lty = par("lty"),
  lwd = par("lwd"),
  code = 3,
  length = 0.05,
  pch = NA,
  cex.pch = par("cex"),
  col.pch = par("fg"),
  bg.pch = par("bg"),
  ...
)
}
\arguments{
\item{from}{coordinates of points \bold{from} which to draw (the lower end
of the error bars). If \code{to} is left to \code{NULL} and \code{from} is a
\eqn{k \times 2 }{k x 2} dimensional matrix, the first column will be
interpreted as \code{from} and the second as \code{to}.}

\item{to}{coordinates of points \bold{to} which to draw (the upper end of
the error bars).}

\item{pos}{numeric, position of the error bars. This will either be the
x-coordinate in case of vertical error bars and the y-coordinate in case of
horizontal error bars.}

\item{mid}{numeric, position of midpoints. Defaults to the mean of
\code{from} and \code{to}.}

\item{horiz}{logical, determining whether horizontal error bars are needed
(default is FALSE).}

\item{col}{the line color.}

\item{lty}{the line type.}

\item{lwd}{line width.}

\item{code}{integer code, determining where end lines are to be drawn.
\code{code = 0} will draw no end lines, \code{code = 1} will draw an end
line on the left (lower) side at (\code{x0[i]}, \code{y0[i]}), \code{code =
2} on the right (upper) side (\code{x1[i]}, \code{y1[i]}) and \code{code =
3} (default) will draw end lines at both ends.}

\item{length}{the length of the end lines.}

\item{pch}{plotting character for the midpoints. The position of the points
is given by \code{mid}. If \code{mid} is left to \code{NULL} the points will
be plotted in the middle of \code{from} and \code{to}. No points will be
plotted if this is set to \code{NA}, which is the default.}

\item{cex.pch}{the character extension for the plotting characters. Default
is \code{par("cex")}.}

\item{col.pch}{the color of the plotting characters. Default is
\code{par("fg")}.}

\item{bg.pch}{the background color of the plotting characters (if pch is set
to 21:25). Default is \code{par("bg")}.}

\item{\dots}{the dots are passed to the \code{\link{arrows}} function.}
}
\description{
Add error bars to an existing plot.
}
\details{
A short wrapper for plotting error bars by means of \code{\link{arrows}}.
}
\examples{


op <- par(no.readonly = TRUE)
par(mfrow=c(2,2))

b <- barplot(1:5, ylim=c(0,6))
errBars(from=1:5-rep(0.5,5), to=1:5+rep(0.5,5), pos=b, length=0.2)

# just on one side
b <- barplot(1:5, ylim=c(0,6))
errBars(from=1:5, to=1:5+rep(0.5,5), pos=b, length=0.2, col="red", code=2, lwd=2)

b <- barplot(1:5, xlim=c(0,6), horiz=TRUE)
errBars(from=1:5, to=1:5+rep(0.2,5), pos=b, horiz=TRUE,  length=0.2, col="red", code=2, lwd=2)

par(xpd=FALSE)
dotchart(1:5, xlim=c(0,6))
errBars(from=1:5-rep(0.2,5), to=1:5+rep(0.2,5), horiz=TRUE, length=0.1)

par(op)

}
\seealso{
\code{\link{arrows}}, \code{\link{lines.loess}}

Other plot.annotation:
\code{\link[=axisBreak]{axisBreak()}},
\code{\link[=barText]{barText()}},
\code{\link[=boxedText]{boxedText()}},
\code{\link[=colorLegend]{colorLegend()}},
\code{\link[=drawBand]{drawBand()}},
\code{\link[=shade]{shade()}},
\code{\link[=stamp]{stamp()}},
\code{\link[=titleRect]{titleRect()}}
}
\concept{confidence-intervals}
\concept{descriptive-statistics}
\concept{graphics}
\concept{plot.annotation}
