% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colToOpaque.R
\name{fade}
\alias{fade}
\title{Fade Colors}
\usage{
fade(col, alpha = 0.5)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{alpha}{Alpha transparency values between 0 and 1.}
}
\value{
Character vector of colors.
}
\description{
Apply transparency and remove the alpha channel afterwards.
}
\seealso{
Other color.manipulation:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=darken]{darken()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}}
}
\concept{color-manipulation}
\concept{color.manipulation}
