% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colToOpaque.R
\name{fade}
\alias{fade}
\title{Fade Colors}
\usage{
fade(col, opacity = 0.5)
}
\arguments{
\item{col}{vector of valid R colors}

\item{opacity}{opacity value between 0 and 1}
}
\value{
Character vector of colors.
}
\description{
Apply transparency and remove the alpha channel afterwards.
}
\seealso{
Other color.manipulation:
\code{\link[=addOpacity]{addOpacity()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=darken]{darken()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}}
}
\concept{color}
\concept{color.manipulation}
\concept{transformation}
