% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grayscale.R
\name{grayscale}
\alias{grayscale}
\title{Convert Colors to Grayscale}
\usage{
grayscale(col)
}
\arguments{
\item{col}{Vector of valid R colors.}
}
\value{
Character vector of grayscale colors.
}
\description{
Convert colors to grayscale using luminance weighting.
}
\details{
Grayscale conversion uses the standard luminance approximation:
\deqn{0.3 R + 0.59 G + 0.11 B}
}
\examples{
op <- par(no.readonly = TRUE)
par(mfcol=c(2,2))

tmp <- 1:3
names(tmp) <- c('red','green','blue')

barplot(tmp, col=c('red','green','blue'))
barplot(tmp, col=grayscale(c('red','green','blue')))

barplot(tmp, col=c('red','#008100','#3636ff'))
barplot(tmp, col=grayscale(c('red','#008100','#3636ff')))

par(op)

}
\seealso{
Other color.manipulation:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=darken]{darken()}},
\code{\link[=fade]{fade()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}}
}
\concept{color-manipulation}
\concept{color.manipulation}
\concept{graphics}
