% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{hexToRGB}
\alias{hexToRGB}
\title{Convert Hex Colors to RGB}
\usage{
hexToRGB(hex)
}
\arguments{
\item{hex}{Character vector of hexadecimal colors.}
}
\value{
Integer matrix with RGB rows.
}
\description{
Convert hexadecimal color strings to an RGB matrix.
}
\seealso{
Other color:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=contrastColor]{contrastColor()}},
\code{\link[=darken]{darken()}},
\code{\link[=findColor]{findColor()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=hcol]{hcol()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}},
\code{\link[=pal]{pal()}},
\code{\link[=palNames]{palNames()}},
\code{\link[=setBackCol]{setBackCol()}}
}
\concept{color}
\concept{color-conversion}
