% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{hexToRGB}
\alias{hexToRGB}
\title{Convert Hex Colors to RGB}
\usage{
hexToRGB(hex)
}
\arguments{
\item{hex}{character vector of hexadecimal colors.}
}
\value{
Integer matrix with RGB rows.
}
\description{
Convert hexadecimal color strings to an RGB matrix.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color}
\concept{color-conversion}
