% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten.R
\name{lighten}
\alias{lighten}
\title{Lighten Colors}
\usage{
lighten(col, amount = 0.2)
}
\arguments{
\item{col}{vector of valid R colors.}

\item{amount}{numeric value between 0 and 1 specifying the
amount of lightening. A value of 0 leaves the color unchanged,
while 1 returns white.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Lighten colors by mixing them with white.
}
\details{
Colors are mixed linearly with white in RGB space:
\deqn{
  x_{new} = x + amount \cdot (255 - x)
}
}
\examples{
lighten("tomato")
lighten(c("tomato", "royalblue"), amount = 0.5)

# increasing amounts
amount <- seq(0, 1, 0.2)
barplot(rep(1, 6), col = sapply(amount, lighten, col = "tomato"),
        names.arg = amount, yaxt = "n", xlab = "amount")

}
\seealso{
Other color.manipulation:
\code{\link[=addOpacity]{addOpacity()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=darken]{darken()}},
\code{\link[=fade]{fade()}},
\code{\link[=mixCol]{mixCol()}}
}
\concept{color}
\concept{color.manipulation}
\concept{transformation}
