% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighten.R
\name{lighten}
\alias{lighten}
\title{Lighten Colors}
\usage{
lighten(col, amount = 0.2)
}
\arguments{
\item{col}{Vector of valid R colors.}

\item{amount}{Numeric value between 0 and 1 specifying the
amount of lightening. A value of 0 leaves the color unchanged,
while 1 returns white.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Lighten colors by mixing them with white.
}
\details{
Colors are mixed linearly with white in RGB space:
\deqn{
  x_{new} = x + amount \cdot (255 - x)
}
}
\seealso{
Other color:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=contrastColor]{contrastColor()}},
\code{\link[=darken]{darken()}},
\code{\link[=findColor]{findColor()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=hcol]{hcol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=mixColors]{mixColors()}},
\code{\link[=pal]{pal()}},
\code{\link[=palNames]{palNames()}},
\code{\link[=setBackCol]{setBackCol()}}
}
\concept{color}
