% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.loess.r
\name{lines.loess}
\alias{lines.loess}
\title{Add a Loess Smoother and Its Confidence Intervals}
\usage{
\method{lines}{loess}(
  x,
  col = Pal()[1],
  lwd = 2,
  lty = "solid",
  type = "l",
  n = 100,
  conf.level = 0.95,
  args.band = NULL,
  ...
)
}
\arguments{
\item{x}{the loess object to be plotted.}

\item{col}{linecolor of the smoother. Default is DescTools's \code{col1}.}

\item{lwd}{line width of the smoother.}

\item{lty}{line type of the smoother.}

\item{type}{type of plot, defaults to \code{"l"}.}

\item{n}{number of points used for plotting the fit.}

\item{conf.level}{confidence level for the confidence interval. Set this to
NA, if no confidence band should be plotted.  Default is 0.95.}

\item{args.band}{list of arguments for the confidence band, such as color or
border (see \code{\link{drawBand}}).}

\item{\dots}{further arguments are passed to the \code{loess()} smoother}
}
\description{
Add a loess smoother to an existing plot. The function first calculates the
prediction of a loess object for a reasonable amount of points, then adds
the line to the plot and inserts a polygon with the confidence intervals.
}
\note{
Loess can result in heavy computational load if there are many points!
}
\examples{

op <- par(no.readonly = TRUE)
par(mfrow=c(1,2))

x <- runif(100)
y <- rnorm(100)
plot(x, y)
lines(loess(y~x))

plot(dist ~ speed, data=cars)
lines(loess(dist ~ speed, data=cars))

plot(dist ~ speed, data=cars)
lines(loess(dist ~ speed, data=cars), conf.level = 0.99,
            args.band = list(col=alpha("red", 0.4), border="black") )

# the default values from scatter.smooth
lines(loess(dist ~ speed, data=cars,
            span=2/3, degree=1, family="symmetric"), col="red")

par(op)

}
\seealso{
\code{\link{loess}}, \code{\link{scatter.smooth}},
\code{\link{smooth.spline}}   \%\%, \code{\link{smoothSpline}}
}
\keyword{aplot}
\keyword{math}
