% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineSep.R
\name{lineSep}
\alias{lineSep}
\title{Create a Line Separator String}
\usage{
lineSep(sep = .getOption("linesep"))
}
\arguments{
\item{sep}{A character string used as separator. If \code{NULL},
a default Unicode box-drawing character (\code{"\u2500"}) is used.
If \code{sep} consists of a single visible character, it is
repeated to match the current console width (minus two characters).}
}
\value{
A character string representing a horizontal separator line.
}
\description{
Generates a character string that can be used as a horizontal line
separator in console output.
}
\details{
ANSI escape sequences (e.g., color codes) are removed using
\code{cli::ansi_strip()} when determining the visible width.

The console width is determined via \code{getOption("width", 80)}.
}
\examples{
lineSep()
lineSep("-")
lineSep(cli::col_red("="))

}
\seealso{
Other string.format: 
\code{\link{fm}()},
\code{\link{fmCI}()},
\code{\link{strAbbr}()},
\code{\link{strAlign}()},
\code{\link{strPad}()},
\code{\link{strTrunc}()},
\code{\link{style}()}
}
\concept{graphics}
\concept{string-formatting}
\concept{string.format}
