% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineToUser.R
\name{lineToUser}
\alias{lineToUser}
\title{Convert Line Coordinates To User Coordinates}
\usage{
lineToUser(line, side)
}
\arguments{
\item{line}{the number of lines}

\item{side}{the side of the plot}
}
\value{
the user coordinates for the given lines
}
\description{
Functions like \code{mtext} or \code{axis} use the \code{line} argument to
set the distance from plot. Sometimes it's useful to have the distance in
user coordinates. \code{lineToUser()} does this nontrivial conversion.
}
\details{
For the \code{lineToUser} function to work, there must be an open plot.
}
\examples{

plot(1:10)
lineToUser(line=2, side=4)

}
\seealso{
\code{\link{mtext}}

Other graphics.utils:
\code{\link[=abcCoords]{abcCoords()}},
\code{\link[=axisBreak]{axisBreak()}},
\code{\link[=barText]{barText()}},
\code{\link[=boxedText]{boxedText()}},
\code{\link[=colLegend]{colLegend()}},
\code{\link[=degToRad]{degToRad()}},
\code{\link[=errBars]{errBars()}},
\code{\link[=lineSep]{lineSep()}},
\code{\link[=lines.lm]{lines.lm()}},
\code{\link[=mar]{mar()}},
\code{\link[=preview]{preview()}},
\code{\link[=spreadOut]{spreadOut()}},
\code{\link[=stamp]{stamp()}},
\code{\link[=textLegend]{textLegend()}},
\code{\link[=titleRect]{titleRect()}},
\code{\link[=unit]{unit()}}
}
\concept{annotation}
\concept{graphics.utils}
