% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{longToRGB}
\alias{longToRGB}
\title{Convert Long Integers to RGB}
\usage{
longToRGB(col)
}
\arguments{
\item{col}{Integer vector.}
}
\value{
RGB matrix.
}
\description{
Decode long integers into RGB values.
}
\seealso{
Other color.conversion:
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToHex]{colToHex()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=rgbToCol]{rgbToCol()}},
\code{\link[=rgbToHex]{rgbToHex()}},
\code{\link[=rgbToLong]{rgbToLong()}}
}
\concept{color-manipulation}
\concept{color.conversion}
