% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{longToRGB}
\alias{longToRGB}
\title{Convert Long Integers to RGB}
\usage{
longToRGB(col)
}
\arguments{
\item{col}{Integer vector.}
}
\value{
RGB matrix.
}
\description{
Decode long integers into RGB values.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color-manipulation}
