% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixCol.R
\name{mixCol}
\alias{mixCol}
\title{Mix Colors}
\usage{
mixCol(col1, col2, weights = 0.5)
}
\arguments{
\item{col1}{first vector of colors.}

\item{col2}{second vector of colors.}

\item{weights}{numeric value between 0 and 1 specifying the
contribution of \code{col2}. A value of 0 returns \code{col1},
while 1 returns \code{col2}.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Mix two sets of colors in RGB space.
}
\details{
Colors are mixed linearly in RGB space:
\deqn{
  x_{mix} = (1 - weights) x_1 + weights x_2
}

All arguments are recycled as necessary.
}
\seealso{
Other color.manipulation:
\code{\link[=addOpacity]{addOpacity()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=darken]{darken()}},
\code{\link[=fade]{fade()}},
\code{\link[=lighten]{lighten()}}
}
\concept{color}
\concept{color.manipulation}
\concept{transformation}
