% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixColors.R
\name{mixColors}
\alias{mixColors}
\title{Mix Colors}
\usage{
mixColors(col1, col2, weights = 0.5)
}
\arguments{
\item{col1}{First vector of colors.}

\item{col2}{Second vector of colors.}

\item{weights}{Numeric value between 0 and 1 specifying the
contribution of \code{col2}. A value of 0 returns \code{col1},
while 1 returns \code{col2}.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Mix two sets of colors in RGB space.
}
\details{
Colors are mixed linearly in RGB space:
\deqn{
  x_{mix} = (1 - weights) x_1 + weights x_2
}

All arguments are recycled as necessary.
}
\seealso{
\code{\link{lighten}},
\code{\link{darken}}

Other color:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=contrastColor]{contrastColor()}},
\code{\link[=darken]{darken()}},
\code{\link[=findColor]{findColor()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=hcol]{hcol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=pal]{pal()}},
\code{\link[=palNames]{palNames()}},
\code{\link[=setBackCol]{setBackCol()}}
}
\concept{color}
