% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{palNames}
\alias{palNames}
\title{List Available Palette Names}
\usage{
palNames(type = c("all", "continuous", "discrete"))
}
\arguments{
\item{type}{character, one of \code{"all"} (default), \code{"continuous"},
or \code{"discrete"}.}
}
\value{
a character vector of palette names.
}
\description{
Returns the names of all palettes available in \code{\link{pal}},
optionally filtered by type.
}
\examples{
palNames()
palNames("continuous")
palNames("discrete")

}
\seealso{
\code{\link{pal}}

Other color:
\code{\link[=addAlpha]{addAlpha()}},
\code{\link[=colToOpaque]{colToOpaque()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=contrastColor]{contrastColor()}},
\code{\link[=darken]{darken()}},
\code{\link[=findColor]{findColor()}},
\code{\link[=grayscale]{grayscale()}},
\code{\link[=hcol]{hcol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=lighten]{lighten()}},
\code{\link[=mixColors]{mixColors()}},
\code{\link[=pal]{pal()}},
\code{\link[=setBackCol]{setBackCol()}}
}
\concept{color}
\concept{palette}
