% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BlandAltman.R
\name{plot.BlandAltman}
\alias{plot.BlandAltman}
\title{Bland-Altman Plot}
\usage{
\method{plot}{BlandAltman}(
  x,
  main = "",
  xlab = "Mean",
  ylab = "Difference",
  xlim = NULL,
  ylim = NULL,
  pch = 16,
  col = NULL,
  grid = TRUE,
  meanLine = TRUE,
  limits = TRUE,
  conf.band = FALSE,
  trend = FALSE,
  showText = TRUE,
  stamp = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"blandAltman"}.}

\item{main}{plot title.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{xlim}{X-axis limits.}

\item{ylim}{Y-axis limits.}

\item{pch}{plotting symbol.}

\item{col}{point colour.}

\item{grid}{logical; draw a background grid.}

\item{meanLine}{logical; draw the bias line.}

\item{limits}{logical; draw the limits of agreement.}

\item{conf.band}{logical; draw confidence bands.}

\item{trend}{logical; draw a regression line of
differences versus means.}

\item{showText}{logical; annotate bias and limits.}

\item{stamp}{optional plot stamp.}

\item{...}{further arguments passed to \code{plot()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Plots a Bland-Altman agreement analysis.
}
\details{
Objects of class \code{"blandAltman"} are typically created with
\code{DescToolsX::blandAltmanData()}.
}
\seealso{
Other plot.s3:
\code{\link[=plot.Desc.qn]{plot.Desc.qn()}},
\code{\link[=plot.Desc.table]{plot.Desc.table()}},
\code{\link[=plot.Lc]{plot.Lc()}}
}
\concept{agreement}
\concept{method-comparison}
\concept{plot.s3}
