% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLc.R
\name{plot.Lc}
\alias{plot.Lc}
\alias{lines.Lc}
\title{Plot and Add Lines for Lorenz Curve Objects}
\usage{
\method{plot}{Lc}(
  x,
  general = FALSE,
  lwd = 2,
  type = "l",
  xlab = "p",
  ylab = "L(p)",
  main = "Lorenz curve",
  las = 1,
  pch = NA,
  ...
)

\method{lines}{Lc}(x, general = FALSE, lwd = 2, conf.level = NA, args.cband = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{"Lc"} as returned by \code{\link[DescToolsX]{lc}}.}

\item{general}{Logical. If \code{TRUE}, the generalised Lorenz curve
is plotted.}

\item{lwd}{Line width.}

\item{type}{Plot type (see \code{\link{plot}}).}

\item{xlab, ylab}{Axis labels.}

\item{main}{Plot title.}

\item{las}{Axis label style.}

\item{pch}{Optional plotting character for points.}

\item{...}{Additional graphical parameters passed to
\code{\link{plot}} or \code{\link{lines}}.}

\item{conf.level}{Confidence level for the confidence band.
If \code{NA}, no band is drawn.}

\item{args.cband}{list of arguments for the confidence band, such as color
or border (see \code{\link{drawBand}}).}
}
\value{
Invisibly returns the input object.
}
\description{
S3 methods for visualising objects of class \code{"Lc"}.
The \code{plot} method draws a Lorenz curve, while
\code{lines} adds a Lorenz curve (optionally with confidence band)
to an existing plot.
}
\seealso{
Other topic.graphics: 
\code{\link{axisBreak}()},
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{canvas}()},
\code{\link{colorLegend}()},
\code{\link{errBars}()},
\code{\link{plotArea}()},
\code{\link{plotBag}()},
\code{\link{plotBar}()},
\code{\link{plotBubble}()},
\code{\link{plotCirc}()},
\code{\link{plotCor}()},
\code{\link{plotDens}()},
\code{\link{plotDens2D}()},
\code{\link{plotDot}()},
\code{\link{plotECDF}()},
\code{\link{plotFdist}()},
\code{\link{plotFun}()},
\code{\link{plotLines}()},
\code{\link{plotMiss}()},
\code{\link{plotMosaic}()},
\code{\link{plotProbDist}()},
\code{\link{plotQQ}()},
\code{\link{plotRidge}()},
\code{\link{plotTernary}()},
\code{\link{plotTimeSeries}()},
\code{\link{plotTreemap}()},
\code{\link{plotViolin}()},
\code{\link{plotWeb}()}
}
\concept{Inequality}
\concept{Lorenz Curve}
\concept{base-graphics}
\concept{plotting}
\concept{topic.graphics}
