% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotArea.R
\name{plotArea}
\alias{plotArea}
\alias{plotArea.default}
\alias{plotArea.formula}
\title{Create an Area Plot}
\usage{
plotArea(x, ...)

\method{plotArea}{default}(
  x,
  y = NULL,
  prop = FALSE,
  add = FALSE,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  frame.plot = FALSE,
  ...
)

\method{plotArea}{formula}(formula, data, subset, na.action, ...)
}
\arguments{
\item{x}{numeric vector of x values, or if \code{y=NULL} a numeric vector of
y values. Can also be a 1-dimensional table (x values in names, y values in
array), matrix or 2-dimensional table (x values in row names and y values in
columns), a data frame (x values in first column and y values in subsequent
columns), or a time-series object of class \code{ts/mts}.}

\item{\dots}{further arguments are passed to \code{matplot} and
\code{polygon}.}

\item{y}{numeric vector of y values, or a matrix containing y values in
columns.}

\item{prop}{whether data should be plotted as proportions, so stacked areas
equal 1.}

\item{add}{whether polygons should be added to an existing plot.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{col}{fill color of polygon(s). The default is a vector of gray colors.}

\item{frame.plot}{a logical indicating whether a box should be drawn around
the plot.}

\item{formula}{a \code{\link{formula}}, such as \code{y ~ x} or
\code{cbind(y1, y2) ~ x}, specifying x and y values. A dot on the left-hand
side, \code{formula = . ~ x}, means all variables except the one specified
on the right-hand side.}

\item{data}{a data frame (or list) from which the variables in
\code{formula} should be taken.}

\item{subset}{an optional vector specifying a subset of observations to be
used.}

\item{na.action}{a function which indicates what should happen when the data
contain \code{NA} values. Defaults to \code{getOption("na.action")}.}
}
\value{
Matrix of cumulative sums that was used for plotting.
}
\description{
Produce a stacked area plot, or add polygons to an existing plot.
}
\examples{
plotArea(VADeaths, xaxt="n")
axis(1, at=1:5, labels=rownames(VADeaths))

# formula
plotArea(Armed.Forces~Year, data=longley)
plotArea(cbind(Armed.Forces,Unemployed)~Year, data=longley)

# add=TRUE
plot(1940:1970, 500*runif(31), ylim=c(0,500))
plotArea(Armed.Forces~Year, data=longley, add=TRUE)

# matrix
plotArea(WorldPhones)
plotArea(WorldPhones, prop=TRUE, col=rainbow(10))

# table
tab <- xtabs(ncases ~ agegp + alcgp, data = esoph)
plotArea(tab, xaxt="n", xlab="age group")
axis(1, labels=rownames(tab), at=seq(nrow(tab)))

# ts/mts
plotArea(austres)
plotArea(Seatbelts[,c("drivers","front","rear")],
         ylab="Killed or seriously injured")
abline(v=1983+1/12, lty=3)

}
\seealso{
\code{\link{barplot}}, \code{\link{polygon}},
\code{\link[areaplot]{areaplot}}

Other plot.special:
\code{\link[=plotBinaryTree]{plotBinaryTree()}},
\code{\link[=plotCirc]{plotCirc()}},
\code{\link[=plotFun]{plotFun()}},
\code{\link[=plotMiss]{plotMiss()}},
\code{\link[=plotPolar]{plotPolar()}},
\code{\link[=plotTernary]{plotTernary()}},
\code{\link[=plotTreemap]{plotTreemap()}},
\code{\link[=plotWeb]{plotWeb()}}
}
\author{
Arni Magnusson \href{mailto:thisisarni@gmail.com}{thisisarni@gmail.com}
}
\concept{frequency-analysis}
\concept{graphics}
\concept{plot.special}
