% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotECDF.R
\name{plotECDF}
\alias{plotECDF}
\title{Empirical Cumulative Distribution Function}
\usage{
plotECDF(
  x,
  breaks = NULL,
  col = pal()[1],
  ylab = "",
  lwd = 2,
  xlab = NULL,
  ...
)
}
\arguments{
\item{x}{numeric vector of the observations for ecdf.}

\item{breaks}{will be passed directly to \code{\link{hist}}. If left to
\code{NULL}, no histogram will be used.}

\item{col}{color of the line.}

\item{ylab}{label for the y-axis.}

\item{lwd}{line width.}

\item{xlab}{label for the x-axis.}

\item{\dots}{arguments to be passed to subsequent functions.}
}
\value{
no value returned, use \code{\link{plot.ecdf}} if any results are
required.
}
\description{
Faster alternative for plotting the empirical cumulative distribution
function (ecdf). The function offers the option to construct the ecdf on the
base of a histogram, which makes sense, when x is large. So the plot process
is much faster, without loosing much precision in the details.
}
\details{
The stats function \code{\link{plot.ecdf}} is fine for vectors that are not
too large. However for n ~ 1e7 we would observe a dramatic performance
breakdown (possibly in combination with the use of \code{\link{do.call}}).

\code{PlotECDF} is designed as alternative for quicker plotting the ecdf for
larger vectors. If \code{breaks} are provided as argument, a histogram with
that number of breaks will be calculated and the ecdf will use those
frequencies instead of respecting every single point. \cr Note that a plot
will rarely need more than ~1'000 points on x to have a sufficient
resolution on usual terms. \code{\link{plotFdist}} will also use this number
of breaks by default.
}
\examples{

plotECDF(faithful$eruptions)

# make large vector
x <- rnorm(n=1e7)

# plot only 1000 points instead of 1e7
plotECDF(x, breaks=1000)
}
\seealso{
\code{\link{plot.ecdf}}, \code{\link{plotFdist}}

Other plot.distribution:
\code{\link[=plotCatDist]{plotCatDist()}},
\code{\link[=plotFdist]{plotFdist()}},
\code{\link[=plotProbDist]{plotProbDist()}},
\code{\link[=plotQQ]{plotQQ()}}
}
\concept{descriptive-statistics}
\concept{graphics}
\concept{nonparametric}
\concept{plot.distribution}
