% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFdist.R
\name{plotFdist}
\alias{plotFdist}
\title{Frequency Distribution Plot}
\usage{
plotFdist(
  x,
  main = deparse(substitute(x)),
  xlab = "",
  xlim = NULL,
  na.rm = FALSE,
  args.hist = NULL,
  args.rug = NA,
  args.dens = NULL,
  args.curve = NA,
  args.boxplot = NULL,
  args.ecdf = NULL,
  args.curve.ecdf = NA,
  heights = NULL,
  pdist = NULL,
  ...
)
}
\arguments{
\item{x}{the numerical variable, whose distribution is to be plotted.}

\item{main}{main title of the plot. Use \code{NA} to suppress the title.}

\item{xlab}{label of the x-axis, defaults to \code{""}. (The name of the
variable is typically placed in the main title and would be redundant here.)}

\item{xlim}{range of the x-axis, defaults to a pretty \code{range(x, na.rm =
TRUE)}.}

\item{na.rm}{logical, should \code{NA}s be omitted? Histogram and boxplot
could do without this option, but the density-function refuses to plot with
missings. Defaults to \code{FALSE}.}

\item{args.hist}{list of additional arguments to be passed to the histogram
\code{hist()}.  The defaults chosen when setting \code{args.hist = NULL} are
more or less the same as in \code{\link{hist}}. The argument \code{type}
defines, whether a histogram (\code{"hist"}) or a plot with \code{type =
"h"} (for 'histogram' like vertical lines for \code{mass} representation)
should be used.  The arguments for a "h-plot"" will be \code{col},
\code{lwd}, \code{pch.col}, \code{pch}, \code{pch.bg} for the line and for
an optional point character on top.  The default type used will be chosen on
the structure of \code{x}. If \code{x} is an integer with up to 12 unique
values there will be a "h-plot" and else a histogram!}

\item{args.rug}{list of additional arguments to be passed to the function
\code{rug()}.  Use \code{args.rug = NA} if no rug should be added. This is
the default. Use \code{args.rug = NULL} to add rug with reasonable default
values.}

\item{args.dens}{list of additional arguments to be passed to
\code{density}.  Use \code{args.dens = NA} if no density curve should be
drawn. The defaults are taken from \code{\link{density}}.}

\item{args.curve}{list of additional arguments to be passed to
\code{\link{curve}}.  This argument allows to add a fitted distribution
curve to the histogram. By default no curve will be added (\code{args.curve
= NA}). If the argument is set to \code{NULL}, a normal curve with
\code{mean(x)} and \code{sd(x)} will be drawn. See examples for more
details.}

\item{args.boxplot}{list of additional arguments to be passed to the boxplot
\code{boxplot()}.  The defaults are pretty much the same as in
\code{\link{boxplot}}.  The two additional arguments \code{pch.mean}
(default \code{23}) and \code{col.meanci} (default \code{"grey80"}) control,
if the mean is displayed within the boxplot. Setting those arguments to
\code{NA} will prevent them from being displayed.}

\item{args.ecdf}{list of additional arguments to be passed to \code{ecdf()}.
Use \code{args.ecdf = NA} if no empirical cumulation function should be
included in the plot.  The defaults are taken from \code{\link{plot.ecdf}}.}

\item{args.curve.ecdf}{list of additional arguments to be passed to
\code{\link{curve}}.  This argument allows to add a fitted distribution
curve to the cumulative distribution function. By default no curve will be
added (\code{args.curve.ecdf = NA}). If the argument is set to \code{NULL},
a normal curve with \code{mean(x)} and \code{sd(x)} will be drawn. See
examples for more details.}

\item{heights}{heights of the plotparts, defaults to \code{c(2,0.5,1.4)} for
the histogram, the boxplot and the empirical cumulative distribution
function, resp. to \code{c(2,1.5)} for a histogram and a boxplot only.}

\item{pdist}{distances of the plotparts, defaults to \code{c(0, 0)}, say
there will be no distance between the histogram, the boxplot and the
ecdf-plot. This can be useful for instance in case that the x-axis has to be
added to the histogram.}

\item{\dots}{can be used to set graphic parameters (\code{\link{par}})
as cex.main, las etc.}
}
\description{
This function was developed to create a univariate graphical representation
of the frequency distribution of a numerical vector.
It combines a histogram, a density
curve, a boxplot and the empirical cumulative distribution function (ecdf)
in one single plot. A rug as well as a model distribution curve (e.g. a
normal curve) can optionally be superposed. This results in a dense and
informative picture of the facts. Still the function remains flexible as all
possible arguments can be passed to the single components (\code{hist},
\code{boxplot} etc.) as a list (see examples).
}
\details{
Performance has been significantly improved, but if \code{x} is growing
large (n > 1e7) the function will take its time to complete. Especially the
density curve and the ecdf, but as well as the boxplot (due to the chosen
alpha channel) will take their time to calculate and plot.\cr In such cases
consider taking a sample, i.e. \code{ plotFdist(x[sample(length(x),
size=5000)])}, the big picture of the distribution won't usually change
much. .
}
\examples{
op <- par(no.readonly=TRUE)

plotFdist(x=faithful$eruptions, na.rm=TRUE)

# define additional arguments for hist, dens and boxplot
# do not display the mean and its CI on the boxplot
plotFdist(faithful$eruptions, args.hist=list(breaks=50),
  args.dens=list(col="olivedrab4"), na.rm=TRUE,
  args.boxplot=list(col="olivedrab2", pch.mean=NA, col.meanci=NA))


# do a "h"-plot instead of a histogram for integers
x <- sample(runif(10), 100, replace = TRUE)
plotFdist(x, args.hist=list(type="mass"))

pp <- rpois(n = 100, lambda = 3)
plotFdist(pp, args.hist = list(type="mass", pch=21, col="darkorange2",
          cex.pch=2.5, col.pch="deeppink4", lwd=3, bg.pch="white"),
          args.boxplot = NULL, args.ecdf = NA, main="Probability mass function")

# special arguments for hist, density and ecdf
plotFdist(x=faithful$eruptions,
          args.hist=list(breaks=20), args.dens=list(bw=.1),
          args.ecdf=list(cex=1.2, pch=16, lwd=1), args.rug=TRUE)

# no density curve, no ecdf but add rug instead, make boxplot a bit higher
plotFdist(x=faithful$eruptions, na.rm=TRUE, args.dens=NA, args.ecdf=NA,
  args.hist=list(xaxt="s"),  # display x-axis on the histogram
  args.rug=TRUE, heights=c(3, 2.5), pdist=2.5, main="Delivery time")

# alpha channel on rug is cool, but takes its time for being drawn...
plotFdist(x=faithful$eruptions, args.rug=list(col=alpha("black", 0.1)), na.rm=TRUE)

# plot a normal density curve, but no boxplot nor ecdf
x <- rnorm(1000)
plotFdist(x, args.curve = NULL, args.boxplot=NA, args.ecdf=NA)

# compare with a t-distribution
plotFdist(x, args.curve = list(expr="dt(x, df=2)", col="darkgreen"),
          args.boxplot=NA, args.ecdf=NA)
legend(x="topright", legend=c("kernel density", "t-distribution (df=2)"),
       fill=c("deeppink4", "darkgreen"), xpd=NA)

# add a gamma distribution curve to both, histogram and ecdf
ozone <- airquality$Ozone
m <- mean(ozone, na.rm = TRUE)
v <- var(ozone, na.rm = TRUE)

plotFdist(ozone, args.hist = list(breaks=15),
  args.curve = list(expr="dgamma(x, shape = m^2/v, scale = v/m)", 
                    col="navajowhite3"),
  args.curve.ecdf = list(expr="pgamma(x, shape = m^2/v, scale = v/m)", 
                         col="navajowhite3"),
  na.rm = TRUE, main = "Airquality - Ozone")

       
par(op)
        
}
\seealso{
\code{\link{hist}}, \code{\link{boxplot}}, \code{\link{ecdf}},
\code{\link{density}}, \code{\link{rug}}, \code{\link{layout}}

Other topic.graphics: 
\code{\link{axisBreak}()},
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{canvas}()},
\code{\link{colorLegend}()},
\code{\link{errBars}()},
\code{\link{plot.Desc.factor}()},
\code{\link{plot.Desc.logical}()},
\code{\link{plot.Desc.table}()},
\code{\link{plot.Desc.ts}()},
\code{\link{plot.Lc}()},
\code{\link{plot.PostHocTest}()},
\code{\link{plotArea}()},
\code{\link{plotBag}()},
\code{\link{plotBar}()},
\code{\link{plotBubble}()},
\code{\link{plotCirc}()},
\code{\link{plotCor}()},
\code{\link{plotDens}()},
\code{\link{plotDens2D}()},
\code{\link{plotDot}()},
\code{\link{plotECDF}()},
\code{\link{plotFun}()},
\code{\link{plotLines}()},
\code{\link{plotMiss}()},
\code{\link{plotMosaic}()},
\code{\link{plotProbDist}()},
\code{\link{plotQQ}()},
\code{\link{plotRidge}()},
\code{\link{plotTernary}()},
\code{\link{plotTimeSeries}()},
\code{\link{plotTreemap}()},
\code{\link{plotViolin}()},
\code{\link{plotWeb}()}
}
\concept{base-graphics}
\concept{plotting}
\concept{topic.graphics}
