% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMiss.R
\name{plotMiss}
\alias{plotMiss}
\title{Plot Missing Data}
\usage{
plotMiss(
  x,
  col = "deeppink4",
  bg = fade("navajowhite3", 0.3),
  clust = FALSE,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{a data.frame to be analysed.}

\item{col}{the colour of the missings.}

\item{bg}{the background colour of the plot.}

\item{clust}{logical, defining if the missings should be clustered. Default
is \code{FALSE}.}

\item{main}{the main title.}

\item{\dots}{the dots are passed to \code{\link{plot}}.}
}
\value{
if clust is set to TRUE, the new order will be returned invisibly.
}
\description{
Takes a data frame and displays the location of missing data. The missings
can be clustered and be displayed together.
}
\details{
A graphical display of the position of the missings can be help to detect
dependencies or patterns within the missings.
}
\note{
Following an idea of Henk Harmsen \href{mailto:henk@carbonmetrics.com}{henk@carbonmetrics.com}
}
\examples{

plotMiss(airquality, main="Missing data (in orignal order)")
plotMiss(airquality, main="Missing data (clustered)", clust=TRUE)

}
\seealso{
\code{\link{hclust}}, \code{\link[bedrock]{countCompCases}}

Other topic.graphics: 
\code{\link{axisBreak}()},
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{canvas}()},
\code{\link{colorLegend}()},
\code{\link{errBars}()},
\code{\link{plot.Lc}()},
\code{\link{plotArea}()},
\code{\link{plotBag}()},
\code{\link{plotBar}()},
\code{\link{plotBubble}()},
\code{\link{plotCirc}()},
\code{\link{plotCor}()},
\code{\link{plotDens}()},
\code{\link{plotDens2D}()},
\code{\link{plotDot}()},
\code{\link{plotECDF}()},
\code{\link{plotFdist}()},
\code{\link{plotFun}()},
\code{\link{plotLines}()},
\code{\link{plotMosaic}()},
\code{\link{plotProbDist}()},
\code{\link{plotQQ}()},
\code{\link{plotRidge}()},
\code{\link{plotTernary}()},
\code{\link{plotTimeSeries}()},
\code{\link{plotTreemap}()},
\code{\link{plotViolin}()},
\code{\link{plotWeb}()}
}
\concept{base-graphics}
\concept{plotting}
\concept{topic.graphics}
