% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMosaic.R
\name{plotMosaic}
\alias{plotMosaic}
\title{Mosaic Plots}
\usage{
plotMosaic(
  x,
  main = deparse(substitute(x)),
  horiz = TRUE,
  cols = NULL,
  off = 0.02,
  mar = NULL,
  xlab = NULL,
  ylab = NULL,
  cex = par("cex"),
  las = 2,
  ...
)
}
\arguments{
\item{x}{a contingency table in array form, with optional category labels
specified in the dimnames(x) attribute. The table is best created by the
table() command. So far only 2-way tables are allowed.}

\item{main}{character string for the mosaic title.}

\item{horiz}{logical, defining the orientation of the mosaicplot.
\code{TRUE} (default) makes a horizontal plot.}

\item{cols}{the colors of the plot.}

\item{off}{the offset between the rectangles. Default is 0.02.}

\item{mar}{the margin for the plot.}

\item{xlab, ylab}{x- and y-axis labels used for the plot; by default, the
first and second element of names(dimnames(X)) (i.e., the name of the first
and second variable in X).}

\item{cex}{numeric character expansion factor; multiplied by
\code{par("cex")} yields the final character size. \code{NULL} and \code{NA}
are equivalent to 1.0.}

\item{las}{the style of axis labels. 0 - parallel to the axis, 1 -
horizontal, 2 - perpendicular, 3 - vertical.}

\item{...}{Graphical parameters passed to \code{par()}, such as
\code{cex}, \code{las}, \code{mar}, or \code{oma}.}
}
\value{
list with the midpoints of the rectangles
}
\description{
Plots a mosaic on the current graphics device.
}
\details{
The reason for this function to exist are the unsatisfying labels in base
mosaicplot.
}
\examples{

plotMosaic(HairEyeColor[,,1])

}
\references{
Friendly, M. (1994) Mosaic displays for multi-way contingency
tables. \emph{Journal of the American Statistical Association}, \bold{89},
190-200.
}
\seealso{
\code{\link{mosaicplot}}
}
\keyword{hplot}
