% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProbDist.R
\name{plotProbDist}
\alias{plotProbDist}
\title{Plot Probability Distribution}
\usage{
plotProbDist(
  breaks,
  FUN,
  blab = NULL,
  main = "",
  xlim = NULL,
  col = NULL,
  density = 7,
  alab = LETTERS[1:(length(breaks) - 1)],
  alab_x = NULL,
  alab_y = NULL,
  ylab = "density",
  ...
)
}
\arguments{
\item{breaks}{a numeric vector containing the breaks of different areas. The
start and end must not be infinity.}

\item{FUN}{the (typically) distribution function}

\item{blab}{text for labelling the breaks}

\item{main}{main title for the plot}

\item{xlim}{the x-limits for the plot}

\item{col}{the color for the shaded areas}

\item{density}{the density for the shaded areas}

\item{alab}{the labels for areas}

\item{alab_x}{the x-coord for the area labels}

\item{alab_y}{the y-coord for the area labels, if left to default they will
be placed in the middle of the plot}

\item{ylab}{the label for they y-axis}

\item{\dots}{further parameters passed to internally used function
\code{\link{curve}()}}
}
\value{
nothing returned
}
\description{
Produce a plot from a probability distribution with shaded areas. This is
often needed in theory texts for classes in statistics.
}
\details{
The function sets up a two-step plot procedure based on curve() and shade()
with additional labelling for convenience.
}
\examples{
# plot t-distribution
plotProbDist(breaks=c(-6, -2.3, 1.5, 6), 
             function(x) dt(x, df=8), 
             blab=c("A","B"), xlim=c(-4,4), alab=NA,
             main="t-Distribution (df=8)",
             col=c("deeppink4", "skyblue3", "darkorange2"), 
             density=c(20, 7))

# Normal
plotProbDist(breaks=c(-10, -1, 12), 
             function(x) dnorm(x, mean=2, sd=2), 
             blab="A", xlim=c(-7,10),
             main="Normal-Distribution N(2,2)",
             col=c("deeppink4", "skyblue3"), 
             density=c(20, 7))

# same for Chi-square
plotProbDist(breaks=c(0, 15, 35), 
             function(x) dchisq(x, df=8), 
             blab="B", xlim=c(0, 30),
             main=expression(paste(chi^2-Distribution, " (df=8)")),
             col=c("deeppink4", "skyblue3"), density=c(0, 20))


}
\seealso{
\code{\link{shade}}, \code{\link{curve}}, \code{\link{polygon}}

Other topic.graphics: 
\code{\link{axisBreak}()},
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{canvas}()},
\code{\link{colorLegend}()},
\code{\link{errBars}()},
\code{\link{plot.Desc.factor}()},
\code{\link{plot.Desc.logical}()},
\code{\link{plot.Desc.table}()},
\code{\link{plot.Desc.ts}()},
\code{\link{plot.Lc}()},
\code{\link{plot.PostHocTest}()},
\code{\link{plotArea}()},
\code{\link{plotBag}()},
\code{\link{plotBar}()},
\code{\link{plotBubble}()},
\code{\link{plotCirc}()},
\code{\link{plotCor}()},
\code{\link{plotDens}()},
\code{\link{plotDens2D}()},
\code{\link{plotDot}()},
\code{\link{plotECDF}()},
\code{\link{plotFdist}()},
\code{\link{plotFun}()},
\code{\link{plotLines}()},
\code{\link{plotMiss}()},
\code{\link{plotMosaic}()},
\code{\link{plotQQ}()},
\code{\link{plotRidge}()},
\code{\link{plotTernary}()},
\code{\link{plotTimeSeries}()},
\code{\link{plotTreemap}()},
\code{\link{plotViolin}()},
\code{\link{plotWeb}()}
}
\concept{base-graphics}
\concept{plotting}
\concept{topic.graphics}
