% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Combined Plot of a Time Series and Its ACF and PACF}
\usage{
plotTimeSeries(
  x,
  lag.max = 10 * log10(length(x)),
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{univariate time series.}

\item{lag.max}{integer. Defines the number of lags to be displayed. The
default is 10 * log10(length(series)).}

\item{ylab}{a title for the y axis: see \code{\link{title}}.}

\item{main}{an overall title for the plot}

\item{\dots}{the dots are passed to the plot command.}
}
\description{
Combined plot of a time Series and its autocorrelation and partial
autocorrelation
}
\details{
plotTimeSeries plots a combination of the time series and its autocorrelation and
partial autocorrelation.
}
\note{
Rewritten based on ideas of M.Huerzeler
}
\examples{

plotTimeSeries(AirPassengers)
}
\seealso{
\code{\link{ts}}

Other topic.graphics: 
\code{\link{axisBreak}()},
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{canvas}()},
\code{\link{colorLegend}()},
\code{\link{errBars}()},
\code{\link{plot.Desc.factor}()},
\code{\link{plot.Desc.logical}()},
\code{\link{plot.Desc.table}()},
\code{\link{plot.Desc.ts}()},
\code{\link{plot.Lc}()},
\code{\link{plot.PostHocTest}()},
\code{\link{plotArea}()},
\code{\link{plotBag}()},
\code{\link{plotBar}()},
\code{\link{plotBubble}()},
\code{\link{plotCirc}()},
\code{\link{plotCor}()},
\code{\link{plotDens}()},
\code{\link{plotDens2D}()},
\code{\link{plotDot}()},
\code{\link{plotECDF}()},
\code{\link{plotFdist}()},
\code{\link{plotFun}()},
\code{\link{plotLines}()},
\code{\link{plotMiss}()},
\code{\link{plotMosaic}()},
\code{\link{plotProbDist}()},
\code{\link{plotQQ}()},
\code{\link{plotRidge}()},
\code{\link{plotTernary}()},
\code{\link{plotTreemap}()},
\code{\link{plotViolin}()},
\code{\link{plotWeb}()}
}
\concept{base-graphics}
\concept{plotting}
\concept{topic.graphics}
