% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Combined Plot of a Time Series and Its ACF and PACF}
\usage{
plotTimeSeries(
  x,
  lag.max = 10 * log10(length(x)),
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{univariate time series.}

\item{lag.max}{integer. Defines the number of lags to be displayed. The
default is 10 * log10(length(series)).}

\item{ylab}{a title for the y axis: see \code{\link{title}}.}

\item{main}{an overall title for the plot}

\item{\dots}{the dots are passed to the plot command.}
}
\description{
Combined plot of a time Series and its autocorrelation and partial
autocorrelation
}
\details{
plotTimeSeries plots a combination of the time series and its autocorrelation and
partial autocorrelation.
}
\note{
Rewritten based on ideas of M.Huerzeler
}
\examples{

plotTimeSeries(AirPassengers)
}
\seealso{
\code{\link{ts}}

Other plot.timeseries: 
\code{\link{plotLines}()}
}
\concept{descriptive-statistics}
\concept{graphics}
\concept{plot.timeseries}
\concept{time-series}
