% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreemap.R
\name{plotTreemap}
\alias{plotTreemap}
\title{Create a Treemap}
\usage{
plotTreemap(
  x,
  grp = NULL,
  labels = NULL,
  text.col = "black",
  col = rainbow(length(x)),
  labels.grp = NULL,
  cex.grp = 3,
  text.col.grp = "black",
  border.grp = "grey50",
  lwd.grp = 5,
  main = "",
  ...
)
}
\arguments{
\item{x}{a vector storing the values to be used to calculate the areas of
rectangles.}

\item{grp}{a vector specifying the group (i.e. country, sector, etc.) to
which each element belongs.}

\item{labels}{a vector specifying the labels.}

\item{text.col}{the text color of the area labels. Default is "black".}

\item{col}{a vector storing the values to be used to calculate the color of
rectangles.}

\item{labels.grp}{a character vector specifying the labels for the groups.}

\item{cex.grp}{the character extension for the group labels. Default is 3.}

\item{text.col.grp}{the text color of the group labels. Default is "black".}

\item{border.grp}{the border color for the group rectangles. Default is
"grey50". Set this to \code{NA} if no special border is desired.}

\item{lwd.grp}{the linewidth of the group borders. Default is 5.}

\item{main}{a title for the plot.}

\item{...}{Additional graphical parameters (via \code{\link[graphics]{par}}).}

\item{cex}{the character extension for the area labels. Default is 1.}
}
\value{
returns a list with groupwise organized midpoints in x and y for the
rectangles within a group and for the groups themselves.
}
\description{
Creates a treemap where rectangular regions of different size, color, and
groupings visualize the elements.
}
\details{
A treemap is a two-dimensional visualization for quickly analyzing large,
hierarchical data sets. Treemaps are unique among visualizations because
they provide users with the ability to see both a high level overview of
data as well as fine-grained details. Users can find outliers, notice
trends, and perform comparisons using treemaps. Each data element contained
in a treemap is represented with a rectangle, or a cell. Treemap cell
arrangement, size, and color are each mapped to an attribute of that
element. Treemap cells can be grouped by common attributes. Within a group,
larger cells are placed towards the bottom left, and smaller cells are
placed at the top right.
}
\note{
Parts of the code contributed by Jeff Enos.
}
\examples{

set.seed(1789)
N <- 20
area <- rlnorm(N)

plotTreemap(x=sort(area, decreasing=TRUE), labels=letters[1:20], 
            col=pal("RedToBlack", 20))


grp <- sample(x=1:3, size=20, replace=TRUE, prob=c(0.2,0.3,0.5))

z <- bedrock::sortX(data.frame(area=area, grp=grp), c("grp","area"), 
           decreasing=c(FALSE,TRUE))
z$col <- alpha(c("steelblue","green","yellow")[z$grp],
                  unlist(lapply(split(z$area, z$grp),
                  function(...) linScale(..., newLow=0.1, newHigh=0.6))))

plotTreemap(x=z$area, grp=z$grp, labels=letters[1:20], col=z$col)


b <- plotTreemap(x=z$area, grp=z$grp, labels=letters[1:20], labels.grp=NA,
                 col=z$col, main="Treemap")

# the function returns the midpoints of the areas
# extract the group midpoints from b
mid <- do.call(rbind, lapply(lapply(b, "[", 1), data.frame))

# and draw some visible text
boxedText(x=mid$grp.x, y=mid$grp.y, labels=LETTERS[1:3], 
          cex=3, border=NA,
          col=alpha("white",0.7) )


}
\seealso{
\code{\link{plotCirc}}, \code{\link{mosaicplot}},
\code{\link{barplot}}

Other plot.special:
\code{\link[=plotArea]{plotArea()}},
\code{\link[=plotBinaryTree]{plotBinaryTree()}},
\code{\link[=plotCirc]{plotCirc()}},
\code{\link[=plotFun]{plotFun()}},
\code{\link[=plotMiss]{plotMiss()}},
\code{\link[=plotPolar]{plotPolar()}},
\code{\link[=plotTernary]{plotTernary()}},
\code{\link[=plotWeb]{plotWeb()}}
}
\concept{frequency-analysis}
\concept{graphics}
\concept{plot.special}
