% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeb.R
\name{plotWeb}
\alias{plotWeb}
\title{Plot a Web of Connected Points}
\usage{
plotWeb(
  m,
  col = NULL,
  lty = NULL,
  lwd = NULL,
  args.legend = NULL,
  pch = 21,
  pt.cex = 2,
  pt.col = "black",
  pt.bg = "darkgrey",
  cex.lab = 1,
  las = 1,
  adj = NULL,
  dist = 0.5,
  ...
)
}
\arguments{
\item{m}{a symmetric matrix of numeric values}

\item{col}{the color for the connecting lines}

\item{lty}{the line type for the connecting lines, the default will be
\code{par("lty")}.}

\item{lwd}{the line widths for the connecting lines. If left to \code{NULL}
it will be linearly scaled between the minimum and maximum value of
\code{m}.}

\item{args.legend}{list of additional arguments to be passed to the
\code{legend} function.  Use \code{args.legend = NA} if no legend should be
added.}

\item{pch}{the plotting symbols appearing in the plot, as a non-negative
numeric vector (see \code{\link{points}}, but unlike there negative values
are omitted) or a vector of 1-character strings, or one multi-character
string.}

\item{pt.cex}{expansion factor(s) for the points.}

\item{pt.col}{the foreground color for the points, corresponding to its
argument \code{col}.}

\item{pt.bg}{the background color for the points, corresponding to its
argument \code{bg}.}

\item{cex.lab}{the character extension for the labels.}

\item{las}{alignment of the labels, 1 means horizontal, 2 radial and 3
vertical.}

\item{adj}{adjustments for the labels. (Left: 0, Right: 1, Mid: 0.5)}

\item{dist}{gives the distance of the labels from the outer circle. Default
is 2.}

\item{\dots}{dots are passed to plot.}
}
\value{
A list of x and y coordinates, giving the coordinates of all the
points drawn, useful for adding other elements to the plot.
}
\description{
This plot can be used to graphically display a correlation matrix by using
the linewidth between the nodes in proportion to the correlation of two
variables. It will place the elements homogenously around a circle and draw
connecting lines between the points.
}
\details{
The function uses the lower triangular matrix of \code{m}, so this is the
order colors, linewidth etc. must be given, when the defaults are to be
overrun.
}
\examples{
op <- par(no.readonly = TRUE)

m <- cor(swiss)
plotWeb(m=m, main="Swiss correlation")


# let's describe only the significant corrs and start with a dataset
# get the p-values
p <- outer(
  (vars <- colnames(mtcars)), vars,
  Vectorize(function(v1, v2)
    cor.test(mtcars[[v1]], mtcars[[v2]], method = "pearson")$p.value
  )
)
dimnames(p) <- list(vars, vars)

# ok, got all the p-values, now replace > 0.05 with NAs
m[p > 0.05] <- NA

# How does that look like now?
# try also: fm(m, na.form = ". ", ldigits=0, digits=3, align = "right")

# plotWeb(m, las=2, cex=1.2)

# define line widths
# plotWeb(m, lwd=abs(m[lower.tri(m)] * 10))

par(op)

}
\seealso{
\code{\link{plotCor}}
}
\keyword{hplot}
\keyword{multivariate}
