% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-methods.R
\name{polygon}
\alias{polygon}
\alias{polygon.ringGeometry}
\alias{polygon.polygonGeometry}
\alias{polygon.geometryCollection}
\alias{polygon.default}
\title{Draw Polygonal Geometries}
\usage{
polygon(x, ...)

\method{polygon}{default}(
  x,
  y = NULL,
  density = NULL,
  angle = 45,
  border = NULL,
  col = NA,
  lty = par("lty"),
  ...,
  fillOddEven = FALSE
)

\method{polygon}{ringGeometry}(x, rule = "evenodd", ...)

\method{polygon}{polygonGeometry}(x, ...)
}
\arguments{
\item{x}{An object to be drawn.}

\item{...}{Further arguments passed to the corresponding method.}

\item{y}{Numeric vector of y-coordinates.}

\item{density}{Density of shading lines.}

\item{angle}{Angle of shading lines in degrees.}

\item{border}{Border colour.}

\item{col}{Fill colour.}

\item{lty}{Line type.}

\item{fillOddEven}{Logical; should the odd-even rule be used for filling?}

\item{rule}{Character string specifying the filling rule passed to
\code{\link[graphics]{polypath}}. One of \code{"evenodd"} or
\code{"winding"}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Generic function for drawing polygon-based geometry objects.
This function extends graphics::polygon() with support for
geometry objects such as circle(), ellipse(), regPolygon() and ring()
and further remains fully compatible with its original interface.
#'
For ordinary coordinate vectors the call is forwarded to
\code{\link[graphics]{polygon}}. Geometry objects such as
\code{\link{circle}}, \code{\link{ellipse}}, \code{\link{regPolygon}}
and \code{\link{ring}} are dispatched to specialised methods.
}
\examples{
canvas()

polygon(
  circle(radius = 1),
  col = "lightblue"
)

polygon(
  regPolygon(
    radius = 0.7,
    numVertices = 5
  ),
  border = "red"
)

}
\seealso{
\code{\link{arc}},
\code{\link{circle}},
\code{\link{ellipse}},
\code{\link{regPolygon}},
\code{\link{ring}},
\code{\link[graphics]{polygon}}

Other geometry:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=canvas]{canvas()}},
\code{\link[=polarGrid]{polarGrid()}},
\code{\link[=rotate]{rotate()}},
\code{\link[=shade]{shade()}},
\code{\link[=transformXY]{transformXY()}}
}
\concept{geometry}
