% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{regPolygon}
\alias{regPolygon}
\title{Regular Polygon Geometry}
\usage{
regPolygon(x = 0, y = 0, radius = 1, nVertices = 6, startAngle = 0)
}
\arguments{
\item{x, y}{centre coordinates.}

\item{radius}{circumradius.}

\item{nVertices}{number of vertices.}

\item{startAngle}{staring angle in radians.}
}
\value{
An object inheriting from class \code{"regPolygonGeometry"}.
}
\description{
Create a regular polygon.
}
\seealso{
Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=circle]{circle()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=polygonX]{polygonX()}},
\code{\link[=ring]{ring()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
