% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{regPolygon}
\alias{regPolygon}
\title{Regular Polygon Geometry}
\usage{
regPolygon(x = 0, y = 0, radius = 1, numVertices = 6, startAngle = 0)
}
\arguments{
\item{x, y}{Centre coordinates.}

\item{radius}{Circumradius.}

\item{numVertices}{Number of vertices.}

\item{startAngle}{Staring angle in radians.}
}
\value{
An object inheriting from class \code{"regPolygonGeometry"}.
}
\description{
Create a regular polygon.
}
