% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{RGBToCMY}
\alias{RGBToCMY}
\title{Convert RGB to CMY}
\usage{
RGBToCMY(col, maxColorValue = 1)
}
\arguments{
\item{col}{RGB matrix or hexadecimal colors.}

\item{maxColorValue}{Maximum channel value.}
}
\value{
Numeric CMY matrix.
}
\description{
Convert RGB colors to the CMY color space.
}
\seealso{
Other color.space:
\code{\link[=CMYKToCMY]{CMYKToCMY()}},
\code{\link[=CMYKToRGB]{CMYKToRGB()}},
\code{\link[=CMYToCMYK]{CMYToCMYK()}}
}
\concept{color-conversion}
\concept{color-space}
\concept{color.space}
\concept{graphics}
