% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{rgbToCmy}
\alias{rgbToCmy}
\title{Convert RGB to CMY}
\usage{
rgbToCmy(col, maxColorValue = 1)
}
\arguments{
\item{col}{RGB matrix or hexadecimal colors.}

\item{maxColorValue}{Maximum channel value.}
}
\value{
Numeric CMY matrix.
}
\description{
Convert RGB colors to the CMY color space.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color-conversion}
\concept{color-space}
\concept{graphics}
