% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{rgbToCol}
\alias{rgbToCol}
\title{Convert RGB Colors to the Nearest Named R Color}
\usage{
rgbToCol(col, method = c("rgb", "hsv"), metric = c("euclidean", "manhattan"))
}
\arguments{
\item{col}{RGB matrix or hexadecimal colors.}

\item{method}{Color space used for matching.}

\item{metric}{Distance metric.}
}
\value{
Character vector of named R colors.
}
\description{
Match RGB colors to the nearest named R color.
}
\seealso{
Other color.conversion:
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToHex]{colToHex()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=longToRGB]{longToRGB()}},
\code{\link[=rgbToHex]{rgbToHex()}},
\code{\link[=rgbToLong]{rgbToLong()}}
}
\concept{color-conversion}
\concept{color.conversion}
\concept{graphics}
