% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{RGBToCol}
\alias{RGBToCol}
\title{Convert RGB Colors to the Nearest Named R Color}
\usage{
RGBToCol(col, method = c("rgb", "hsv"), metric = c("euclidean", "manhattan"))
}
\arguments{
\item{col}{RGB matrix or hexadecimal colors.}

\item{method}{Color space used for matching.}

\item{metric}{Distance metric.}
}
\value{
Character vector of named R colors.
}
\description{
Match RGB colors to the nearest named R color.
}
\seealso{
Other color.conversion:
\code{\link[=RGBToHex]{RGBToHex()}},
\code{\link[=RGBToLong]{RGBToLong()}},
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToHex]{colToHex()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=longToRGB]{longToRGB()}}
}
\concept{color-manipulation}
\concept{color.conversion}
