% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{rgbToHex}
\alias{rgbToHex}
\title{Convert RGB to Hexadecimal Colors}
\usage{
rgbToHex(col)
}
\arguments{
\item{col}{RGB matrix.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Convert RGB values to hexadecimal color strings.
}
\seealso{
Other color.conversion:
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToHex]{colToHex()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=longToRGB]{longToRGB()}},
\code{\link[=rgbToCol]{rgbToCol()}},
\code{\link[=rgbToLong]{rgbToLong()}}
}
\concept{color-conversion}
\concept{color.conversion}
\concept{graphics}
