% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{RGBToHex}
\alias{RGBToHex}
\title{Convert RGB to Hexadecimal Colors}
\usage{
RGBToHex(col)
}
\arguments{
\item{col}{RGB matrix.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Convert RGB values to hexadecimal color strings.
}
\seealso{
Other color.conversion:
\code{\link[=RGBToCol]{RGBToCol()}},
\code{\link[=RGBToLong]{RGBToLong()}},
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToHex]{colToHex()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=longToRGB]{longToRGB()}}
}
\concept{color-manipulation}
\concept{color.conversion}
