% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{rgbToHex}
\alias{rgbToHex}
\title{Convert RGB to Hexadecimal Colors}
\usage{
rgbToHex(col)
}
\arguments{
\item{col}{RGB matrix.}
}
\value{
Character vector of hexadecimal colors.
}
\description{
Convert RGB values to hexadecimal color strings.
}
\seealso{
\link{color-conversion-overview}
}
\concept{color-conversion}
\concept{graphics}
