% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{rgbToLong}
\alias{rgbToLong}
\title{Convert RGB to Long Integers}
\usage{
rgbToLong(col)
}
\arguments{
\item{col}{RGB matrix.}
}
\value{
Integer vector.
}
\description{
Encode RGB colors as long integers.
}
\seealso{
Other color.conversion:
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToHex]{colToHex()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=longToRGB]{longToRGB()}},
\code{\link[=rgbToCol]{rgbToCol()}},
\code{\link[=rgbToHex]{rgbToHex()}}
}
\concept{color-conversion}
\concept{color.conversion}
\concept{graphics}
