% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-conversions.R
\name{RGBToLong}
\alias{RGBToLong}
\title{Convert RGB to Long Integers}
\usage{
RGBToLong(col)
}
\arguments{
\item{col}{RGB matrix.}
}
\value{
Integer vector.
}
\description{
Encode RGB colors as long integers.
}
\seealso{
Other color.conversion:
\code{\link[=RGBToCol]{RGBToCol()}},
\code{\link[=RGBToHex]{RGBToHex()}},
\code{\link[=colToHSV]{colToHSV()}},
\code{\link[=colToHex]{colToHex()}},
\code{\link[=colToRGB]{colToRGB()}},
\code{\link[=hexToCol]{hexToCol()}},
\code{\link[=hexToRGB]{hexToRGB()}},
\code{\link[=longToRGB]{longToRGB()}}
}
\concept{color-conversion}
\concept{color.conversion}
\concept{graphics}
