% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{ring}
\alias{ring}
\title{Ring Geometry}
\usage{
ring(
  x = 0,
  y = 0,
  innerRadius = 0.5,
  outerRadius = 1,
  startAngle = 0,
  endAngle = 2 * pi,
  nPoints = 100
)
}
\arguments{
\item{x, y}{centre coordinates.}

\item{innerRadius}{radius of the inner boundary.}

\item{outerRadius}{radius of the outer boundary.}

\item{startAngle, endAngle}{start and end angle in radians.}

\item{nPoints}{number of points used for each boundary.}
}
\value{
An object inheriting from class \code{"ringGeometry"} or a
\code{"geometryCollection"}.
}
\description{
Create one or more rings or ring segments.
}
\seealso{
Other geometry.structures:
\code{\link[=arc]{arc()}},
\code{\link[=band]{band()}},
\code{\link[=bezier]{bezier()}},
\code{\link[=circle]{circle()}},
\code{\link[=ellipse]{ellipse()}},
\code{\link[=polygonX]{polygonX()}},
\code{\link[=regPolygon]{regPolygon()}}
}
\concept{geometry}
\concept{geometry.structures}
\concept{numerical-methods}
