% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry-constructors.R
\name{ring}
\alias{ring}
\title{Ring Geometry}
\usage{
ring(
  x = 0,
  y = 0,
  innerRadius = 0.5,
  outerRadius = 1,
  startAngle = 0,
  endAngle = 2 * pi,
  numPoints = 100
)
}
\arguments{
\item{x, y}{Centre coordinates.}

\item{innerRadius}{Radius of the inner boundary.}

\item{outerRadius}{Radius of the outer boundary.}

\item{startAngle, endAngle}{Start and end angle in radians.}

\item{numPoints}{Number of points used for each boundary.}
}
\value{
An object inheriting from class \code{"ringGeometry"} or a
\code{"geometryCollection"}.
}
\description{
Create one or more rings or ring segments.
}
