% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate a Geometric Structure}
\usage{
rotate(x, y = NULL, mx = NULL, my = NULL, theta = pi/3, asp = 1)
}
\arguments{
\item{x, y}{vectors containing the coordinates of the vertices of the polygon
, which has to be rotated.  The coordinates can be passed in a plotting
structure (a list with x and y components), a two-column matrix, .... See
\code{\link[=xy.coords]{xy.coords()}}.}

\item{mx, my}{xy-coordinates of the center of the rotation. If left to NULL,
the centroid of the structure will be used.}

\item{theta}{angle of the rotation}

\item{asp}{the aspect ratio for the rotation. Helpful for rotate structures
along an ellipse.}
}
\value{
The function invisibly returns a list of the coordinates for the
rotated shape(s).
}
\description{
Rotate a geometric structure by an angle theta around a centerpoint xy.
}
\examples{
op <- par(no.readonly = TRUE)
# let's have a triangle
canvas(main="Rotation")
x <- regPolygon(numVertices=3)

# and rotate
sapply( (0:3) * pi/6, function(theta) {
  xy <- rotate( x=x, theta=theta )
  polygon(xy, col=addOpacity("blue", 0.2))
} )

abline(v=0, h=0)

par(op)

}
\seealso{
\code{\link[=polygon]{polygon()}}, \code{\link[=regPolygon]{regPolygon()}},
\code{\link[=ellipse]{ellipse()}}, \code{\link[=arc]{arc()}}

Other geometry.transformation:
\code{\link[=transformXY]{transformXY()}}
}
\concept{geometry}
\concept{geometry.transformation}
\concept{numerical-methods}
