% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBackCol.R
\name{setBackCol}
\alias{setBackCol}
\title{Background of a Plot}
\usage{
setBackCol(col = "grey", region = c("plot", "figure"), border = NA)
}
\arguments{
\item{col}{the color of the background, if two colors are provided, the
first is used for the plot region and the second for the figure region.}

\item{region}{either \code{"plot"} or \code{"figure"}}

\item{border}{color for rectangle border(s). Default is \code{NA} for no
borders.}
}
\description{
Paints the background of the plot, using either the figure region, the plot
region or both. It can sometimes be cumbersome to elaborate the coordinates
and base R does not provide a simple function for that.
}
\examples{

# use two different colors for the figure region and the plot region
plot(x = rnorm(100), col="blue", cex=1.2, pch=16,
     panel.first={setBackCol(c("red", "lightyellow"))
                  grid()})


}
\seealso{
\code{\link[=rect]{rect()}}

Other graphics.setup:
\code{\link[=canvas]{canvas()}},
\code{\link[=polarGrid]{polarGrid()}}
}
\concept{color}
\concept{geometry}
\concept{graphics.setup}
