% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade.R
\name{shade}
\alias{shade}
\title{Sometimes the area under a density curve has to be color shaded, for
instance to illustrate a p-value or a specific region under the normal
curve. This function draws a curve corresponding to a function over the
interval \code{[from, to]}. It can plot also an expression in the variable
\code{xname}, default \code{x}.}
\usage{
shade(expr, col = par("fg"), breaks, density = 10, n = 101, xname = "x", ...)
}
\arguments{
\item{expr}{the name of a function, or a \code{\link{call}} or an
\code{\link{expression}} written as a function of \code{x} which will
evaluate to an object of the same length as \code{x}.}

\item{col}{color to fill or shade the shape with. The default is taken from
\code{par("fg")}.}

\item{breaks}{numeric, a vector giving the breakpoints between the distinct
areas to be shaded differently. Should be finite as there are no plots with
infinite limits.}

\item{density}{the density of the lines as needed in polygon.}

\item{n}{integer; the number of x values at which to evaluate. Default is
101.}

\item{xname}{character string giving the name to be used for the x axis.}

\item{\dots}{the dots are passed on to \code{\link{polygon}}.}
}
\value{
A list with components \code{x} and \code{y} of the points that were
drawn is returned invisibly.
}
\description{
Useful for shading the area under a curve as often needed for explaining
significance tests.
}
\examples{

curve(dt(x, df=5), xlim=c(-6,6),
      main=paste("Student t-Distribution Probability Density Function, df = ", 5, ")", sep=""),
      type="n", las=1, ylab="probability", xlab="t")

shade(dt(x, df=5), breaks=c(-6, qt(0.025, df=5), qt(0.975, df=5), 6),
      col=c("deeppink4", "skyblue3"), density=c(20, 7))

}
\seealso{
\code{\link{polygon}}, \code{\link{curve}}

Other plot.annotation: 
\code{\link{axisBreak}()},
\code{\link{barText}()},
\code{\link{boxedText}()},
\code{\link{colorLegend}()},
\code{\link{drawBand}()},
\code{\link{errBars}()},
\code{\link{stamp}()},
\code{\link{titleRect}()}
}
\concept{descriptive-statistics}
\concept{distributions}
\concept{graphics}
\concept{plot.annotation}
\keyword{aplot}
