% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.smooth.spline.R
\name{splineCI}
\alias{splineCI}
\alias{smoothSpline}
\alias{lines.smooth.spline}
\alias{lines.SmoothSpline}
\alias{smoothSpline.default}
\alias{smoothSpline.formula}
\title{Add a Spline Smoother}
\usage{
smoothSpline(x, ...)

\method{smoothSpline}{default}(x, ...)

\method{smoothSpline}{formula}(formula, data, subset, na.action, weights, ...)

\method{lines}{SmoothSpline}(
  x,
  col = pal()[1],
  lwd = 2,
  lty = "solid",
  type = "l",
  conf.level = 0.95,
  args.band = NULL,
  ...
)

\method{lines}{smooth.spline}(
  x,
  col = pal()[1],
  lwd = 2,
  lty = "solid",
  type = "l",
  conf.level = 0.95,
  args.band = NULL,
  ...
)
}
\arguments{
\item{x}{the smooth.spline object to be plotted.}

\item{\dots}{further arguments are passed to the smoother \code{SmoothSpline()}).}

\item{formula}{A formula of the form \code{lhs ~ rhs}, where \code{lhs}
gives the response values and \code{rhs} the corresponding groups
or explanatory variables.}

\item{data}{An optional matrix or data frame (or similar; see
\code{\link[stats]{model.frame}}) containing the variables in the
formula. By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{An optional vector specifying a subset of observations
to be used in the analysis.}

\item{na.action}{A function which indicates what should happen when
the data contain \code{NA}s. Defaults to
\code{getOption("na.action")}.}

\item{weights}{optional vector of weights of the same length as x; defaults to all 1.}

\item{col}{linecolor of the smoother. Default is DescToolsX's \code{col1}.}

\item{lwd}{line width of the smoother.}

\item{lty}{line type of the smoother.}

\item{type}{type of plot, defaults to \code{"l"}.}

\item{conf.level}{confidence level for the confidence interval. Set this to
NA, if no confidence band should be plotted.  Default is 0.95.}

\item{args.band}{list of arguments for the confidence band, such as color or
border (see \code{\link{drawBand}}).}
}
\description{
Add a spline smoother to an existing plot. The function first calculates the
prediction of a spline object for a reasonable amount of points, then adds
the line to the plot and inserts a polygon with the confidence intervals.
}
\examples{

op <- par(no.readonly = TRUE)
par(mfrow=c(1,2))

x <- runif(100)
y <- rnorm(100)
plot(x, y)
lines(smooth.spline(y ~ x))

plot(dist ~ speed, data=cars)
lines(smoothSpline(dist ~ speed, data=cars))

par(op)

}
\seealso{
\code{\link{loess}}, \code{\link{scatter.smooth}}

Other plot.utils:
\code{\link{axTicks}},
\code{\link[=lineToUser]{lineToUser()}},
\code{\link[=lines.lm]{lines.lm()}},
\code{\link[=mar]{mar()}},
\code{\link[=setBackCol]{setBackCol()}},
\code{\link[=spreadOut]{spreadOut()}}
}
\concept{graphics}
\concept{plot.utils}
\concept{regression}
\keyword{aplot}
\keyword{math}
