% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lines.smooth.spline.R
\name{splineCI}
\alias{splineCI}
\alias{splineX}
\alias{lines.splineX}
\alias{lines.SplineX}
\alias{splineX.default}
\alias{splineX.formula}
\title{Add a Spline Smoother}
\usage{
splineX(x, ...)

\method{splineX}{default}(x, ...)

\method{splineX}{formula}(formula, data, subset, na.action = na.omit, weights, ...)

\method{lines}{SplineX}(
  x,
  col = pal()[1],
  lwd = 2,
  lty = "solid",
  type = "l",
  bandArgs = list(conf.level = 0.95),
  ...
)
}
\arguments{
\item{x}{spline object returned by \code{splineX()}.}

\item{\dots}{further arguments passed to
\code{\link[stats:smooth.spline]{smooth.spline}}.}

\item{formula}{A formula of the form \code{lhs ~ rhs}, where \code{lhs}
gives the response values and \code{rhs} the corresponding groups
or explanatory variables.}

\item{data}{An optional matrix or data frame (or similar; see
\code{\link[stats]{model.frame}}) containing the variables in the
formula. By default the variables are taken from
\code{environment(formula)}.}

\item{subset}{An optional vector specifying a subset of observations
to be used in the analysis.}

\item{na.action}{A function which indicates what should happen when
the data contain \code{NA}s. Defaults to
\code{getOption("na.action")}.}

\item{weights}{optional vector of weights of the same length as x.}

\item{col}{line color of the smoother.}

\item{lwd}{line width.}

\item{lty}{line type.}

\item{type}{plotting type passed to \code{\link{lines}}.}

\item{bandArgs}{controls the confidence band. May be \code{TRUE},
\code{FALSE}, \code{NULL}, \code{NA}, or a named list. The confidence
level is specified via \code{conf.level}. Default is
\code{list(conf.level = 0.95)}.}
}
\description{
Fit a smoothing spline and optionally add confidence bands.
}
\details{
Confidence bands are controlled via \code{bandArgs}. These arguments can be:
\itemize{
\item \code{FALSE}, \code{NULL} or \code{NA}: suppress the band
\item \code{TRUE}: draw the band with default settings
\item a named list: customize the band appearance and confidence level
}
}
\examples{
op <- par(no.readonly = TRUE)
par(mfrow = c(1, 2))

x <- runif(100)
y <- rnorm(100)

plot(x, y)
lines(splineX(y ~ x))

plot(dist ~ speed, cars)
lines(splineX(dist ~ speed, cars))

plot(dist ~ speed, cars)
lines(
  splineX(dist ~ speed, cars),
  bandArgs = list(
    conf.level = 0.99,
    col = addAlpha("red", 0.3),
    border = "black"
  )
)

par(op)

}
\seealso{
\code{\link{loess}}, \code{\link{scatter.smooth}}

Other plot.utils:
\code{\link{axTicks}}
}
\concept{graphics}
\concept{plot.utils}
\concept{regression}
\keyword{aplot}
\keyword{math}
