% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\title{Date/Time/Directory Stamp the Current Plot}
\usage{
stamp(txt = NULL, las = NULL, cex = 0.6, col = "grey40")
}
\arguments{
\item{txt}{an optional single text string. If it is not given, the function
will look for a defined option named \code{stamp}. If not found the current
date will be taken as text. If the stamp option is defined as expression the
function will evaluate it. This can be used to define dynamic texts. Use \code{NA}
to not display a stamp at all.}

\item{las}{numeric in \code{c(1, 3)}, defining direction of the text. 1
means horizontal, 3 vertical. Default is taken from \code{par("las")}.}

\item{cex}{numeric \bold{c}haracter \bold{ex}pansion factor; multiplied by
\code{par("cex")} yields the final character size. Defaults to 0.6.}

\item{col}{the text color}
}
\description{
Stamp the current plot in the extreme lower right corner. A free text or
expression can be defined as text to the stamp.
}
\details{
The text can be freely defined as option. If user and date should be
included by default, the following option using an expression will help:
\preformatted{setDescToolsXOption(stamp=expression(gettextf('%s/%s',
Sys.getenv('USERNAME'), Today() )))}

For \R results may not be satisfactory if \code{par(mfrow=)} is in effect.
}
\examples{

plot(1:20)
stamp()
}
\seealso{
\code{\link{text}}

Other plot.annotation:
\code{\link[=axisBreak]{axisBreak()}},
\code{\link[=barText]{barText()}},
\code{\link[=boxedText]{boxedText()}},
\code{\link[=colorLegend]{colorLegend()}},
\code{\link[=drawBand]{drawBand()}},
\code{\link[=errBars]{errBars()}},
\code{\link[=shade]{shade()}},
\code{\link[=titleRect]{titleRect()}}
}
\concept{date-handling}
\concept{graphics}
\concept{plot.annotation}
\keyword{aplot}
