% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp.R
\name{stamp}
\alias{stamp}
\title{Date/Time/Directory Stamp the Current Plot}
\usage{
stamp(text = .useTheme, las = NULL, cex = 0.6, col = "grey40")
}
\arguments{
\item{text}{Character string, expression, or toggle controlling the
stamp text. \code{.useTheme} (default) or \code{TRUE} resolve to
\code{getTheme()$stamp}, evaluated lazily at draw time. \code{FALSE},
\code{NULL}, or \code{NA} suppress the stamp. Any other string or
unevaluated \code{expression()} is used as given.}

\item{las}{orientation; see \code{\link[graphics]{par}}. \code{NULL}
(default) inherits the current \code{par("las")}. \code{las = 3}
places the stamp vertically along the right edge instead of
horizontally along the bottom.}

\item{cex, col}{size and color of the stamp text.}
}
\description{
Stamp the current plot in the extreme lower right corner. A free text or
expression can be defined as text to the stamp.
}
\details{
The text can be freely defined as option. If user and date should be
included by default, the following option using an expression will help:
\preformatted{setDescToolsXOption(stamp=expression(gettextf('%s/%s',
Sys.getenv('USERNAME'), Today() )))}

For \R results may not be satisfactory if \code{par(mfrow=)} is in effect.
}
\examples{

plot(1:20)
stamp()

}
\seealso{
\code{\link{text}}

Other graphics.utils:
\code{\link[=abcCoords]{abcCoords()}},
\code{\link[=axisBreak]{axisBreak()}},
\code{\link[=barText]{barText()}},
\code{\link[=boxedText]{boxedText()}},
\code{\link[=colLegend]{colLegend()}},
\code{\link[=degToRad]{degToRad()}},
\code{\link[=errBars]{errBars()}},
\code{\link[=lineSep]{lineSep()}},
\code{\link[=lineToUser]{lineToUser()}},
\code{\link[=lines.lm]{lines.lm()}},
\code{\link[=mar]{mar()}},
\code{\link[=preview]{preview()}},
\code{\link[=spreadOut]{spreadOut()}},
\code{\link[=textLegend]{textLegend()}},
\code{\link[=titleRect]{titleRect()}},
\code{\link[=unit]{unit()}}
}
\concept{annotation}
\concept{graphics.utils}
\keyword{aplot}
