% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strCountW.R
\name{strCountW}
\alias{strCountW}
\title{Count Words in Strings}
\usage{
strCountW(x)
}
\arguments{
\item{x}{a character vector}
}
\value{
An integer vector giving the number of words in each element of
\code{x}.
}
\description{
Counts the number of words in each element of a character vector.
}
\details{
Words are detected using Unicode-aware word boundaries as implemented in
\pkg{stringi}. This ensures robust handling of different languages,
punctuation, and whitespace.
}
\examples{
strCountW("This is a sentence.")

strCountW(c("One word", "Two words here", NA))

}
\seealso{
\code{\link[stringi]{stri_count_words}}

\link{string-overview} for an overview of all string utilities in lyra.
}
\concept{string-inspection}
\concept{summary}
