% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strDist.R
\name{strDist}
\alias{strDist}
\title{Compute Distances Between Strings}
\usage{
strDist(
  x,
  y,
  method = "levenshtein",
  mismatch = 1,
  gap = 1,
  ignoreCase = FALSE
)
}
\arguments{
\item{x}{character vector, first string.}

\item{y}{character vector, second string.}

\item{method}{character, name of the distance method. This must be
\code{"levenshtein"}, \code{"normlevenshtein"} or \code{"hamming"}. Default
is \code{"levenshtein"}, the classical Levenshtein distance.}

\item{mismatch}{numeric, distance value for a mismatch between symbols.}

\item{gap}{numeric, distance value for inserting a gap.}

\item{ignoreCase}{if \code{FALSE} (default), the distance measure will be
case sensitive and if \code{TRUE}, case is ignored.}
}
\value{
\code{strDist} returns an object of class \code{"dist"}; cf.
\code{\link[stats]{dist}}.
}
\description{
\code{strDist} computes distances between strings following to Levenshtein
or Hamming method.
}
\details{
The function computes the Hamming and the Levenshtein (edit) distance of two
given strings (sequences). The Hamming distance between two vectors is the
number mismatches between corresponding entries.

In case of the Hamming distance the two strings must have the same length.

In case of the Levenshtein (edit) distance a scoring and a trace-back matrix
are computed and are saved as attributes \code{"ScoringMatrix"} and
\code{"TraceBackMatrix"}.  The numbers in the trace-back matrix reflect
insertion of a gap in string \code{y} (1), match/missmatch (2), and
insertion of a gap in string \code{x} (3).

The edit distance is useful, but normalizing the distance to fall within the
interval \verb{[0,1]} is preferred because it is somewhat diffcult to judge whether
an LD of for example 4 suggests a high or low degree of similarity.  The
method \code{"normlevenshtein"} for normalizing the LD is sensitive to this
scenario. In this implementation, the Levenshtein distance is transformed to
fall in this interval as follows: \deqn{lnd = 1 - \frac{ld}{max(length(x),
length(y))}}{lnd = 1 - ld / max(length(x), length(y))}

where \code{ld} is the edit distance and \code{max(length(x), length(y))}
denotes that we divide by the length of the larger of the two character
strings. This normalization, referred to as the Levenshtein normalized
distance (lnd), yields a statistic where 1 indicates perfect agreement
between the two strings, and a 0 denotes imperfect agreement. The closer a
value is to 1, the more certain we can be that the character strings are the
same; the closer to 0, the less certain.
}
\note{
For distances between strings and for string alignments see also
Bioconductor package \pkg{Biostrings}
}
\examples{

x <- "GACGGATTATG"
y <- "GATCGGAATAG"
## Levenshtein distance
d <- strDist(x, y)
d
attr(d, "ScoringMatrix")
attr(d, "TraceBackMatrix")

## Hamming distance
strDist(x, y, method="hamming")

}
\references{
R. Merkl and S. Waack (2009) \emph{Bioinformatik Interaktiv}.
Wiley.

Harold C. Doran (2010) \emph{MiscPsycho. An R Package for Miscellaneous
Psychometric Analyses}
}
\seealso{
\code{\link[utils]{adist}}, \code{\link[stats]{dist}}

Other string:
\code{\link[=mgsub]{mgsub()}},
\code{\link[=strAbbr]{strAbbr()}},
\code{\link[=strAlign]{strAlign()}},
\code{\link[=strCap]{strCap()}},
\code{\link[=strChop]{strChop()}},
\code{\link[=strCountW]{strCountW()}},
\code{\link[=strExtract]{strExtract()}},
\code{\link[=strExtractBetween]{strExtractBetween()}},
\code{\link[=strIsNumeric]{strIsNumeric()}},
\code{\link[=strLen]{strLen()}},
\code{\link[=strPad]{strPad()}},
\code{\link[=strPos]{strPos()}},
\code{\link[=strRev]{strRev()}},
\code{\link[=strSpell]{strSpell()}},
\code{\link[=strSplit]{strSplit()}},
\code{\link[=strTrim]{strTrim()}},
\code{\link[=strTrunc]{strTrunc()}},
\code{\link[=strVal]{strVal()}}
}
\author{
Matthias Kohl \href{mailto:Matthias.Kohl@stamats.de}{Matthias.Kohl@stamats.de}
}
\concept{string}
\concept{string-manipulation}
