% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-overview.R
\name{string-overview}
\alias{string-overview}
\title{String Functions in pharos}
\description{
pharos provides a family of functions for manipulating and inspecting
character strings, built on top of \pkg{stringi} for Unicode-aware
behavior. They fall into two groups:

\strong{Manipulation} -- transform, extract, or reshape string content:\tabular{ll}{
   Function \tab Purpose \cr
   \code{\link[=strAbbr]{strAbbr()}} \tab Abbreviate strings uniquely \cr
   \code{\link[=strAlign]{strAlign()}} \tab Align strings \cr
   \code{\link[=strCap]{strCap()}} \tab Capitalize strings \cr
   \code{\link[=strChop]{strChop()}} \tab Split a string into a number of sections of defined length \cr
   \code{\link[=strExtract]{strExtract()}} \tab Extract first match from strings \cr
   \code{\link[=strExtractBetween]{strExtractBetween()}} \tab Extract substrings between patterns \cr
   \code{\link[=strLeft]{strLeft()}} / \code{\link[=strRight]{strRight()}} \tab Return the left or the right part of a string \cr
   \code{\link[=strPad]{strPad()}} \tab Pad a string with justification \cr
   \code{\link[=strRev]{strRev()}} \tab Reverse strings \cr
   \code{\link[=strSpell]{strSpell()}} \tab Spell strings using phonetic alphabets \cr
   \code{\link[=strSplit]{strSplit()}} \tab Split strings \cr
   \code{\link[=strTrim]{strTrim()}} \tab Remove leading/trailing whitespace from a string \cr
   \code{\link[=strTrunc]{strTrunc()}} \tab Truncate strings and add ellipses if a string is truncated \cr
   \code{\link[=strVal]{strVal()}} \tab Extract numeric values from strings \cr
}


\strong{Information} -- inspect properties of strings without changing them:\tabular{ll}{
   Function \tab Purpose \cr
   \code{\link[=strCountW]{strCountW()}} \tab Count words in strings \cr
   \code{\link[=strDist]{strDist()}} \tab Compute distances between strings \cr
   \code{\link[=strIsNumeric]{strIsNumeric()}} \tab Check if character strings represent numeric values \cr
   \code{\link[=strLen]{strLen()}} \tab String length \cr
   \code{\link[=strPos]{strPos()}} \tab Find position of first occurrence of a string \cr
}
}
\seealso{
\code{\link[base:substr]{base::substr()}}
}
