% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strLen.R
\name{strLen}
\alias{strLen}
\title{String length}
\usage{
strLen(x, ...)
}
\arguments{
\item{x}{a character vector}

\item{...}{further arguments passed to \code{nchar}}
}
\description{
Intuitive alias for \code{\link{nchar}}.
}
