% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strPos.R
\name{strPos}
\alias{strPos}
\title{Find Position of First Occurrence Of a String}
\usage{
strPos(x, pattern, pos = 1)
}
\arguments{
\item{x}{a character vector in which to search for the pattern, or an object
which can be coerced by as.character to a character vector.}

\item{pattern}{character string (search string) containing the pattern to be
matched in the given character vector. This can be a character string or a
regular expression.}

\item{pos}{integer, defining the start position for the search within x. The
result will then be relative to the begin of the truncated string. Will be
recycled.}
}
\value{
a vector of the first position of pattern in x
}
\description{
Returns the numeric position of the first occurrence of a substring within a
string. If the search string is not found, the result will be \code{NA}.
}
\details{
This is just a wrapper for the function \code{\link{regexpr}}.
}
\examples{

strPos(x=rownames(mtcars), pattern="y")

# first t, starting with position 4
strPos(x=rownames(mtcars), pattern="t", pos=4)


}
\seealso{
\code{\link{strChop}}, \code{\link{regexpr}}

Other string.manipulation:
\code{\link[=strCap]{strCap()}},
\code{\link[=strChop]{strChop()}},
\code{\link[=strCountW]{strCountW()}},
\code{\link[=strDist]{strDist()}},
\code{\link[=strExtract]{strExtract()}},
\code{\link[=strExtractBetween]{strExtractBetween()}},
\code{\link[=strIsNumeric]{strIsNumeric()}},
\code{\link{strLeftRight}},
\code{\link[=strRev]{strRev()}},
\code{\link[=strSpell]{strSpell()}},
\code{\link[=strSplit]{strSplit()}},
\code{\link[=strTrim]{strTrim()}}
}
\concept{data-inspection}
\concept{string-manipulation}
\concept{string.manipulation}
