% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strRev.R
\name{strRev}
\alias{strRev}
\title{Reverse Strings}
\usage{
strRev(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with each string reversed.
}
\description{
Reverses each element of a character vector.
}
\details{
This function uses \code{\link[stringi]{stri_reverse}}, which correctly
handles Unicode characters and multi-byte encodings.
}
\examples{
strRev("abc")

strRev(c("hello", "world"))

# Unicode-safe
strRev("äöü")

}
\seealso{
\code{\link[stringi]{stri_reverse}}

Other string.manipulation:
\code{\link[=strCap]{strCap()}},
\code{\link[=strChop]{strChop()}},
\code{\link[=strCountW]{strCountW()}},
\code{\link[=strDist]{strDist()}},
\code{\link[=strExtract]{strExtract()}},
\code{\link[=strExtractBetween]{strExtractBetween()}},
\code{\link[=strIsNumeric]{strIsNumeric()}},
\code{\link{strLeftRight}},
\code{\link[=strPos]{strPos()}},
\code{\link[=strSpell]{strSpell()}},
\code{\link[=strSplit]{strSplit()}},
\code{\link[=strTrim]{strTrim()}}
}
\concept{data-manipulation}
\concept{string-manipulation}
\concept{string.manipulation}
