% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strSplit.R
\name{strSplit}
\alias{strSplit}
\title{Split Strings}
\usage{
strSplit(x, split = "", fixed = FALSE)
}
\arguments{
\item{x}{A character vector to be split.}

\item{split}{A character string specifying the delimiter (if
\code{fixed = TRUE}) or a regular expression (if \code{fixed = FALSE}).}

\item{fixed}{Logical; if \code{TRUE}, \code{split} is treated as a fixed
string. Otherwise, it is interpreted as a regular expression.}
}
\value{
A list of character vectors. If \code{x} has length 1, a character
vector is returned.
}
\description{
Splits character vectors into substrings. This is a convenience wrapper
around \code{\link[stringi]{stri_split_fixed}} and
\code{\link[stringi]{stri_split_regex}} with simplified defaults.
}
\details{
If the input \code{x} has length 1, the result is returned as a character
vector instead of a list for convenience.

This function provides a simplified interface to string splitting using
the \pkg{stringi} package. It avoids some of the complexity of
\code{\link{strSplit}} while providing consistent and Unicode-aware
behavior.
}
\examples{
strSplit("a,b,c", ",", fixed = TRUE)

strSplit(c("a b", "c d"), " ")

# regex splitting
strSplit("a1b2c3", "\\\\d")

}
\seealso{
\code{\link[stringi]{stri_split_fixed}},
\code{\link[stringi]{stri_split_regex}}

Other string.manipulation:
\code{\link[=strCap]{strCap()}},
\code{\link[=strChop]{strChop()}},
\code{\link[=strCountW]{strCountW()}},
\code{\link[=strDist]{strDist()}},
\code{\link[=strExtract]{strExtract()}},
\code{\link[=strExtractBetween]{strExtractBetween()}},
\code{\link[=strIsNumeric]{strIsNumeric()}},
\code{\link{strLeftRight}},
\code{\link[=strPos]{strPos()}},
\code{\link[=strRev]{strRev()}},
\code{\link[=strSpell]{strSpell()}},
\code{\link[=strTrim]{strTrim()}},
\code{\link[=strVal]{strVal()}}
}
\concept{data-manipulation}
\concept{string-manipulation}
\concept{string.manipulation}
