% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strTrunc.R
\name{strTrunc}
\alias{strTrunc}
\title{Truncate Strings and Add Ellipses If a String is Truncated.}
\usage{
strTrunc(x, maxlen = 20, ellipsis = "...", wbound = FALSE)
}
\arguments{
\item{x}{a vector of strings.}

\item{maxlen}{the maximum length of the returned strings (NOT counting the
appended ellipsis). \code{maxlen} is recycled.}

\item{ellipsis}{the string to be appended, if the string is longer than the
given maximal length. The default is \code{"..."}.}

\item{wbound}{logical. Determines if the maximal length should be reduced to
the next smaller word boundary and so words are not chopped. Default is
\code{FALSE}.}
}
\value{
The string(s) passed as \samp{x} now with a maximum length of
\samp{maxlen} + 3 (for the ellipsis).
}
\description{
Truncates one or more strings to a specified length, adding an ellipsis
(...)  to those strings that have been truncated. The truncation can also be
performed using word boundaries.  Use \code{\link{strAlign}()} to justify
the strings if needed.
}
\examples{

x <- c("this is short", "and this is a longer text", 
       "whereas this is a much longer story, which could not be told shorter")

# simple truncation on 10 characters
strTrunc(x, maxlen=10)

# NAs remain NA
strTrunc(c(x, NA_character_), maxlen=15, wbound=TRUE)

# using word boundaries
for(i in 0:20)
  print(strTrunc(x, maxlen=i, wbound=TRUE))

# compare
for(i in 0:20)
  print(strTrunc(x, maxlen=i, wbound=FALSE))

}
\seealso{
Other string utilities: 
\code{\link{strAbbr}()},
\code{\link{strAlign}()},
\code{\link{strCap}()},
\code{\link{strCountW}()},
\code{\link{strExtract}()},
\code{\link{strExtractBetween}()},
\code{\link{strPad}()},
\code{\link{strRev}()},
\code{\link{strSpell}()},
\code{\link{strSplit}()},
\code{\link{strTrim}()},
\code{\link{strVal}()}
}
\concept{character data}
\concept{string splitting}
\concept{string utilities}
