% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformXY.R
\name{transformXY}
\alias{transformXY}
\title{Apply Geometric Transformations to Coordinates}
\usage{
transformXY(
  x,
  y = NULL,
  translate = c(0, 0),
  scale = c(1, 1),
  theta = 0,
  asp = 1
)
}
\arguments{
\item{x}{Numeric vector of x coordinates, or an object coercible by
\code{\link{xy.coords}}.}

\item{y}{Numeric vector of y coordinates. Ignored if \code{x} already
contains both coordinates.}

\item{translate}{Numeric vector of length 1 or 2 specifying translation
in x and y direction. Recycled if necessary. Default is \code{c(0, 0)}.}

\item{scale}{Numeric vector of length 1 or 2 specifying scaling factors
for x and y. Recycled if necessary. Default is \code{c(1, 1)}.}

\item{theta}{Rotation angle in radians. Default is \code{0}.}

\item{asp}{Aspect ratio adjustment passed to \code{\link{rotate}}.
Default is \code{1}.}
}
\value{
A list with components \code{x} and \code{y}, as returned by
\code{\link{xy.coords}}.
}
\description{
Applies scaling, rotation, and translation to a set of 2D coordinates.
The transformations are applied in the following order:
\enumerate{
\item Scaling
\item Rotation (see \code{\link{rotate}})
\item Translation
}
}
\details{
This function is a convenience wrapper combining basic affine transformations.
Internally, it uses \code{\link{rotate}} for rotation.
}
\examples{
x <- c(0, 1, 1, 0)
y <- c(0, 0, 1, 1)

# scale, rotate and translate a square
transformXY(x, y,
            scale = 2,
            theta = pi / 4,
            translate = c(1, 1))

# matrix input
m <- cbind(x, y)
transformXY(m, scale = 0.5, theta = pi/6)

}
\seealso{
\code{\link{rotate}}

Other coordinate.transform:
\code{\link{coordinate-conversions}},
\code{\link[=degToRad]{degToRad()}},
\code{\link[=rotate]{rotate()}}
}
\concept{coordinate.transform}
\concept{data-manipulation}
\concept{geometry}
\concept{mathematics}
